% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{delete_prefix}
\alias{delete_prefix}
\title{Delete Prefix}
\usage{
delete_prefix(words)
}
\arguments{
\item{words}{A character vector of words}
}
\value{
A character vector of words with a common prefix
}
\description{
Delete Prefix
}
\examples{
delete_prefix(c("unhappy", "unfair"))

}
