% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{get_prefix}
\alias{get_prefix}
\title{Get the common prefix of words}
\usage{
get_prefix(words)
}
\arguments{
\item{words}{A character vector of words}
}
\value{
A character string that is the common prefix of all words
}
\description{
Get the common prefix of words
}
\examples{
get_prefix(c("animal", "antelope", "anteater"))

}
