% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{make_se_parse}
\alias{make_se_parse}
\title{Title}
\usage{
make_se_parse(
  proteins_unique,
  columns,
  mode = c("char", "delim"),
  chars = 1,
  sep = "_"
)
}
\arguments{
\item{sep}{}
}
\description{
Title
}
