% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{meanSdPlot}
\alias{meanSdPlot}
\title{Plot row standard deviations versus row means}
\usage{
meanSdPlot(
  x,
  ranks = TRUE,
  xlab = ifelse(ranks, "rank(mean)", "mean"),
  ylab = "sd",
  pch,
  plot = TRUE,
  bins = 50,
  ...
)
}
\arguments{
\item{x}{A SummarizedExperiment object.}

\item{ranks}{A logical value indicating whether to plot the rank of the mean instead of the mean.}

\item{xlab}{A character string for the x-axis label.}

\item{ylab}{A character string for the y-axis label.}

\item{pch}{A vector of plotting characters.}

\item{plot}{A logical value indicating whether to plot the mean-sd plot.}

\item{bins}{An integer indicating the number of bins in the plot.}

\item{...}{Other arguments to be passed to the underlying plotting function.}
}
\value{
If plot is TRUE, a mean-sd plot is returned. In any case returns a named list with five components: its elements px and py are the x- and y-coordinates of the individual data points in the plot; its first and second element are the x-coordinates and values of the running median estimator (the red line in the plot). Its element gg is the plot object (see examples). Depending on the value of plot, the method can (and by default does) have a side effect, which is to print gg on the active graphics device.
}
\description{
This function creates a mean-sd plot for a given SummarizedExperiment object.
}
\details{
See \code{\link[vsn]{meanSdPlot}} for further details
}
