% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.ANOVA.Anal}
\alias{met.ANOVA.Anal}
\title{Perform ANOVA analysis}
\usage{
met.ANOVA.Anal(
  mSetObj = NA,
  nonpar = FALSE,
  thresh = 0.05,
  post.hoc = "fisher",
  all_results = FALSE,
  silent = FALSE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object. (see \code{\link[VisomX]{met.read_data}})}

\item{nonpar}{(Logical) Use a non-parametric test (\code{TRUE}) or not (\code{FALSE}).}

\item{thresh}{(Numeric) From 0 to 1, indicate the p-value threshold.}

\item{post.hoc}{(Character) Enter the name of the post-hoc test, "fisher" or "tukey".}

\item{all_results}{(Logical) If \code{TRUE}, it the ANOVA results for all compounds with no post-hoc tests performed will be written as CSV file "anova_all_results.csv".}

\item{silent}{(Logical) Suppress message with number of significant features found (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
The input mSet object with ANOVA results added at mSetObj$analSet$aov.
}
\description{
ANOVA analysis
}
