% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.FC.Anal}
\alias{met.FC.Anal}
\title{Fold change analysis, unpaired}
\usage{
met.FC.Anal(
  mSetObj = NA,
  log2fc.thresh = 1,
  grp1 = NULL,
  grp2 = NULL,
  paired = FALSE
)
}
\arguments{
\item{mSetObj}{Enter the name of the created mSet object. (see \code{\link[VisomX]{met.read_data}})}

\item{log2fc.thresh}{(Numeric) Log2(fold-change) threshold that is considered relevant.}

\item{grp1}{(Character) Enter name of the first group for the contrast \code{grp1 vs. grp2}. If both group arguments are empty, the first two names in the list of groups are selected.}

\item{grp2}{(Character) Enter name of the second group for the contrast \code{grp1 vs. grp2}. If both group arguments are empty, the first two names in the list of groups are selected.}

\item{paired}{(Logical) Are the data in both groups paired (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
The input mSet object with results of fold-change analysis added at mSetObj$analSet$fc$\code{grp1}_vs_\code{grp2}.
}
\description{
Perform fold change analysis, method can be mean or median
}
\references{
adapted from \code{\link[MetaboAnalystR]{FC.Anal}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
