% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.GetFC}
\alias{met.GetFC}
\title{Used by higher functions to calculate fold change}
\usage{
met.GetFC(mSetObj = NA, paired = FALSE, grp1, grp2)
}
\arguments{
\item{mSetObj}{Enter the name of the created mSet object. (see \code{\link[VisomX]{met.read_data}})}

\item{paired}{(Logical) Are the data in both groups paired (\code{TRUE}) or not (\code{FALSE}).}

\item{grp1}{(Character) Enter name of the first group for the contrast \code{grp1 vs. grp2}. If both group arguments are empty, the first two names in the list of groups are selected.}

\item{grp2}{(Character) Enter name of the second group for the contrast \code{grp1 vs. grp2}. If both group arguments are empty, the first two names in the list of groups are selected.}
}
\value{
The input mSet object with results of fold-change analysis added at mSetObj$analSet$fc$\code{grp1}_vs_\code{grp2}.
}
\description{
Utility method to calculate FC, used in higher function
}
\references{
adapted from \code{\link[MetaboAnalystR]{GetFC}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
