% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.GetTtestRes}
\alias{met.GetTtestRes}
\title{Used by higher functions to retrieve T-test p-values}
\usage{
met.GetTtestRes(
  mSetObj = NA,
  grp1,
  grp2,
  paired = FALSE,
  equal.var = TRUE,
  nonpar = F
)
}
\arguments{
\item{mSetObj}{Enter the name of the created mSet object (see \code{\link[VisomX]{met.read_data}}).}

\item{grp1}{(Character) Enter name of the first group for the contrast \code{grp1 vs. grp2}.}

\item{grp2}{(Character) Enter name of the second group for the contrast \code{grp1 vs. grp2}.}

\item{paired}{(Logical) Is the data paired (\code{TRUE}) or not (\code{FALSE}).}

\item{equal.var}{(Logical) Is the group variance equal (\code{TRUE}) or not (\code{FALSE}).}

\item{nonpar}{(Logical) Use a non-parametric test (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A data frame with T-test results.
}
\description{
Utility method to to get p values via T-test.
}
\references{
adapted from \code{\link[MetaboAnalystR]{GetTtestRes}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
