% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.impute}
\alias{met.impute}
\title{Impute missing variables}
\usage{
met.impute(mSetObj = NA, method = "min")
}
\arguments{
\item{mSetObj}{Input name of the created mSet object ((see \code{\link[VisomX]{$met.initialize}} and \code{\link[MetaboAnalystR]{Read.TextData}}).}

\item{method}{(Character) Select the option to replace missing variables:
\itemize{
\item \code{"lod"} replaces missing values with 1/5 of the minimum value for the respective variable.
\item \code{"rowmin"} replaces missing values with the half sample minimum.
\item \code{"colmin"} replaces missing values with the half feature minimum.
\item \code{"mean"} replaces missing values with the mean value of the respective feature column.
\item \code{"median"} replaces missing values with the median value of the respective feature column.
\item \code{"knn_var"} imputes missing values by finding the features in the training set “closest” to it and averages these nearby points to fill in the value.
\item \code{"knn_smp"} imputes missing values by finding the samples in the training set “closest” to it and averages these nearby points to fill in the value.
\item \code{"bpca"} applies Bayesian PCA to impute missing values.
\item \code{"ppca"} applies probabilistic PCA to impute missing values.
\item \code{"svdImpute"} applies singular value decomposition to impute missing values.
}}
}
\value{
The input mSet object with imputed data at mSetObj$dataSet$data_proc.
}
\description{
Replace missing variables via a chosen method. Data need to be re-calibrated after this step, including \code{\link[VisomX::met.PerformFeatureFilter]{filtering}} as well as \code{\link[VisomX::met.normalize]{normalization, transformation, and scaling}}. Data imputation is performed as part of \code{\link[VisomX]{met.read_data}}.
}
