% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.initialize}
\alias{met.initialize}
\title{Constructs a dataSet object for storing metabolomics data}
\usage{
met.initialize(data.type = "conc", anal.type = "stat", paired = FALSE)
}
\arguments{
\item{data.type}{(Character) The type of data, either "list" (Compound lists), "conc" (Compound concentration data), "specbin" (Binned spectra data), "pktable" (Peak intensity table), "nmrpeak" (NMR peak lists), "mspeak" (MS peak lists), or "msspec" (MS spectra data).}

\item{anal.type}{(Character) Indicate the analysis module to be performed: "stat", "pathora", "pathqea", "msetora", "msetssp", "msetqea", "ts", "cmpdmap", "smpmap", or "pathinteg"}

\item{paired}{(Logical) Indicate if the data is paired (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
The input mSet object with imputed data at mSetObj$dataSet$data_proc.
}
\description{
This functions handles the construction of an mSetObj object for storing data for further processing and analysis. It is necessary to utilize this function to provide downstream functions with information about type of data and the type of analysis you will perform, as well as to provide the required data structure. This initialization is performed as part of \code{\link[VisomX]{met.read_data}}.
}
