% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.normalize}
\alias{met.normalize}
\title{Data normalization}
\usage{
met.normalize(
  mSetObj = NA,
  rowNorm = NULL,
  transNorm = NULL,
  scaleNorm = NULL,
  ref = NULL,
  norm.vec = NULL,
  ratio = FALSE,
  ratioNum = 20
)
}
\arguments{
\item{mSetObj}{Enter the name of the created mSet object (see \code{\link[VisomX]{met.read_data}}).}

\item{rowNorm}{(Character) Select the option for row-wise normalization:
\itemize{
\item \code{"GroupPQN"} for probabilistic quotient normalization by a reference group
\item \code{"SamplePQN"} for probabilistic quotient normalization by a reference sample
\item \code{"QuantileNorm"} for Quantile Normalization
\item \code{"CompNorm"} for normalization by a reference feature
\item \code{"SumNorm"} for normalization to constant sum of intensities
\item \code{"MedianNorm"} for normalization to sample median
\item \code{"SpecNorm"} for normalization by a sample-specific factor
}}

\item{transNorm}{(Character) Select option to transform the data:
\itemize{
\item \code{"LogNorm"} for Log10 normalization
\item \code{"CrNorm"} Cubic Root Transformation
}}

\item{scaleNorm}{(Character) Select option for scaling the data:
\itemize{
\item \code{"MeanCenter"} for Mean Centering
\item \code{"AutoNorm"} for Autoscaling
\item \code{"ParetoNorm"} for Pareto Scaling
\item \code{"RangeNorm"} for Range Scaling
}}

\item{ref}{(Character) Enter the name of the reference sample or the reference feature (if \code{rowNorm = "GroupPQN"}, \code{"SamplePQN"}, or \code{"CompNorm"}.}

\item{norm.vec}{(Numeric vector) Vector with sample-specific scaling factors. Only applicable for \code{rowNorm = "SpecNorm"}.}

\item{ratio}{This option is only for biomarker analysis.}

\item{ratioNum}{Relevant only for biomarker analysis.}
}
\value{
The input mSet object with normalized data at mSetObj$dataSet$norm.
}
\description{
\code{met.normalize} performs row-wise normalization, transformation, and scaling of metabolomics data. This step is performed as part of the \code{\link[VisomX]{met.workflow}} function. Additionally, the workflow \code{\link[VisomX]{met.test_normalization}} allows the simultaneous testing of different data processing conditions and helps with finding the most suitable options.
}
\references{
adapted from \code{\link[MetaboAnalystR]{Normalization}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
