% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.PCA.Anal}
\alias{met.PCA.Anal}
\title{Perform a Principal Component Analysis}
\usage{
met.PCA.Anal(mSetObj = NA)
}
\arguments{
\item{mSetObj}{An mSet object containing normalized data}
}
\value{
An mSet object with the PCA results stored in the analSet element.
}
\description{
This function performs a Principal Component Analysis (PCA) on the normalized data set stored in the mSetObj object.
}
