% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.PerformFeatureFilter}
\alias{met.PerformFeatureFilter}
\title{Used by higher functions for non-specific filtering of variables}
\usage{
met.PerformFeatureFilter(
  int.mat,
  mSetObj,
  filter,
  remain.num = remain.num,
  anal.type = NULL,
  all.rsd = NULL
)
}
\arguments{
\item{int.mat}{Data matrix with samples in rows and features in columns, generated by higher function.}

\item{mSetObj}{Input name of the created mSet object (see \code{\link[MetaboAnalystR]{InitDataObjects}} and \code{\link[MetaboAnalystR]{Read.TextData}}).}

\item{filter}{(Character) Select the filter option. \code{"none"} applies no filtering based on the following ranking criteria:
\itemize{
\item \code{"rsd"} filters features with low relative standard deviation across the data set.
\item \code{"nrsd}" is the non-parametric relative standard deviation.
\item \code{"mean"} filters features with low mean intensity value across the data set.
\item \code{"median"} filters features with low median intensity value across the data set.
\item \code{"sd"} filters features with low absolute standard deviation across the data set.
\item \code{"mad"} filters features with low median absolute deviation across the data set.
\item \code{"iqr"} filters features with a low inter-quartile range across the data set.
}}

\item{remain.num}{(Numerical) Enter the number of variables to keep in your data set. If \code{NULL}, the following empirical rules are applied during data filtering with the methods specified in \code{filter = ""}:
\itemize{
\item \strong{Less than 250 variables:} 5\% will be filtered
\item \strong{250 - 500 variables:} 10\% will be filtered
\item \strong{500 - 1000 variables:} 25\% will be filtered
\item \strong{More than 1000 variables:} 40\% will be filtered
}}

\item{anal.type}{(Character) Type of analysis. Extracted from mSetObj by higher function at mSetObj$analSet$type.}

\item{all.rsd}{(Numeric or \code{NULL}) Apply a filter based on the in-group relative standard deviation (RSD, in \%) or not \code{NULL}. Therefore, the RSD of every feature is calculated for every group in the data set. If the RSD of a variable in any group exceeds the indicated threshold, it is removed from the data set. This filter can be applied in addition to other filtering methods and is especially useful to perform on data with technical replicates.}
}
\value{
A list with filtered data and a message to inform about the chosen filtering conditions
}
\description{
Utility method to perform data filtering, used in higher function
}
