% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_CorrHeatMap_Samples}
\alias{met.plot_CorrHeatMap_Samples}
\title{Pattern hunter, plot correlation heatmap}
\usage{
met.plot_CorrHeatMap_Samples(
  mSetObj = NA,
  imgName = "correlation_samples",
  format = "pdf",
  dpi = 72,
  width = NA,
  cor.method = "pearson",
  colors = "bwm",
  viewOpt = "overview",
  fix.col = FALSE,
  no.clst = FALSE,
  corrCutoff = 0,
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object.}

\item{imgName}{(Character) Input a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 for PNG, 72 for PDF)}

\item{width}{(Numeric) width of the the image file in inch (if \code{export = TRUE})?}

\item{cor.method}{Indicate the correlation method, 'pearson', 'spearman', or 'kendall'.}

\item{colors}{Indicate the colors for the heatmap, "bwm" for default, "gbr" for red/green, "heat" for heat colors, "topo" for topo colors, and "gray" for gray scale.}

\item{viewOpt}{Indicate "overview" to get an overview of the heatmap, and "detail" to get a detailed view of the heatmap.}

\item{fix.col}{Logical, fix colors (\code{TRUE}) or not (\code{FALSE}).}

\item{no.clst}{Logical, indicate if the correlations should be clustered (\code{TRUE}) or not (\code{FALSE}).}

\item{corrCutoff}{set correlation cut off}

\item{plot}{(\code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(\code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}
}
\value{
A heat map showing the correlations between samples. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$corr.heatmap_samples.plot)}.
}
\description{
\code{met.plot_CorrHeatMap_Samples} visualizes the correlations between samples.
}
