% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_detect}
\alias{met.plot_detect}
\title{Visualize intensities of compounds with missing values}
\usage{
met.plot_detect(
  mSetObj,
  plot = TRUE,
  imgName = "Plots/DensMissPlot",
  format = "pdf",
  export = TRUE,
  basesize = 10
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container with missing values.}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF and PNG file?}

\item{basesize}{(Numeric) Font size used in the plot.}
}
\value{
Density and CumSum plots of intensities of proteins with and without missing values. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$missval_density.plot)}.
(generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{met.plot_detect} generates density and CumSum plots
of compound intensities with and without missing values.
}
\references{
adapted from package DEP (\url{https://bioconductor.org/packages/DEP/}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
}
