% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_FeatureNormSummary}
\alias{met.plot_FeatureNormSummary}
\title{Visualize effect of normalization on feature density distributions.}
\usage{
met.plot_FeatureNormSummary(
  mSetObj = NA,
  imgName = "FeatureNormSummary",
  format = "pdf",
  dpi = NULL,
  pre.inx = NULL,
  width = NA,
  show_prenorm = TRUE,
  export = TRUE,
  plot = TRUE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object.
Data container after normalization (\code{\link[VisomX]{met.normalize}}).}

\item{imgName}{(Character) Input a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 for PNG, 72 for PDF)}

\item{pre.inx}{Provide index of compounds in your data set that shall be diplayed with box plots. If \code{NULL}, the compounds with minimum and maximum values as well as up to 50 additional features are selected automatically.}

\item{width}{(Numeric) width of the the image file in inch (if \code{export = TRUE})?}

\item{show_prenorm}{(\code{TRUE} or \code{FALSE}) Shall density distributions before normalization be displayed?}

\item{export}{(\code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}

\item{plot}{(\code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}
}
\value{
The top is a density plot and the bottom is a box plot. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$summary_norm_feature.plot)}.
}
\description{
\code{met.plot_FeatureNormSummary} visualizes the density distributions of features before and after normalization, transformation, and scaling.
}
