% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_heatmap}
\alias{met.plot_heatmap}
\title{Create heat map of hierarchically clustered features and samples}
\usage{
met.plot_heatmap(
  mSetObj = NA,
  imgName = "Heatmap_features",
  format = "pdf",
  dpi = NULL,
  width = NA,
  dataOpt = "norm",
  scaleOpt = "row",
  smplDist = "euclidean",
  clstDist = "ward.D",
  palette = "bwm",
  viewOpt = "detail",
  rowV = T,
  colV = T,
  var.inx = NULL,
  border = T,
  grp.ave = F,
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object.}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}).}

\item{format}{(\code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 dpi for PNG, 72 dpi for PDF)}

\item{width}{(Numeric) width of the the image file in inches (if \code{export = TRUE}).}

\item{dataOpt}{(Character) Set data options. Choose \code{"norm"} to use data after normalization or any other string to use original values.}

\item{scaleOpt}{(Character) Set value standardization option. \code{"row"} for autoscaling of features, \code{"column"} for autoscaling of samples, \code{"none"} for no autoscaling.}

\item{smplDist}{(Character) Enter the sample distance method: \code{"euclidean"} for Euclidean distance, \code{"correlation"} for Pearson distance, or \code{"minkowski"} for Minkowski method.}

\item{clstDist}{(Character) Enter the feature clustering method: \code{"ward.d"}, \code{"average"}, \code{"complete"}, or \code{"single"}.}

\item{palette}{Input color palette choice: \code{"bwm"} for Blue-White-Red, \code{"gbr"} for Green-Black-Red, \code{"heat"} for Red-Yellow, \code{"topo"} for topological colors, \code{"gray"} for gray-scale.}

\item{viewOpt}{Set view options, choose \code{"detail"} or \code{"overview"}.}

\item{rowV}{(Logical) Shall clustering be applied to samples?}

\item{colV}{(Logical) Shall clustering be applied to features?}

\item{var.inx}{(Numeric vector) \emph{optional}: Provide indeces of selected features to be plotted.}

\item{border}{(Logical) Indicate whether or not to show cell-borders.}

\item{grp.ave}{(Logical) Shall only group averages be shown (as opposed to displaying every sample)?}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}
}
\value{
The input mSet object with added heat map. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$heatmap.avg.plot)} or \code{print(mSetObj$imgSet$heatmap.all.plot)}, depending on \code{grp.ave}.
}
\description{
\code{met.plot_heatmap} plots a heat map with optional filtering for top features based on results from t-tests/ANOVA, VIP or randomforest.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PlotHeatMap}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
}
