% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_ImpVar}
\alias{met.plot_ImpVar}
\title{Plot PLS important features}
\usage{
met.plot_ImpVar(
  mSetObj = NA,
  imp.vec,
  xlbl,
  feat.num = 15,
  color.BW = FALSE,
  title = FALSE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container after partial least squares-discriminant analysis (\code{\link[VisomX]{met.PLSR.Anal}} and \code{\link[VisomX]{met.PLSDA.CV}}).}

\item{imp.vec}{A vector of importan variables.}

\item{xlbl}{The x-label}

\item{feat.num}{(Numeric) Indicate the number of features to show in the plot.}

\item{color.BW}{(Logical) \code{TRUE} to use black and white, or \code{FALSE} to not.}

\item{title}{(Logical) \code{TRUE} to add a title with the used normalization, transformation, and scaling method, or \code{FALSE} to not add any title.}
}
\value{
The input mSet object with added plot object.
}
\description{
\code{met.plot_ImpVar} is an internal function used by \code{met.plot_PLS_Imp} to visualize the features' importance in the PLS-DA model.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PlotImpVar}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
