% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_missval}
\alias{met.plot_missval}
\title{Plot a heatmap of compounds with missing values}
\usage{
met.plot_missval(
  mSetObj,
  plot = TRUE,
  imgName = "MissValPlot",
  format = "pdf",
  export = TRUE,
  fontsize = 12
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container with missing values.}

\item{plot}{(\code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{imgName}{(Character) Input a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{export}{(\code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF and PNG file?}

\item{fontsize}{(Numeric) Font sized used in the plot.}
}
\value{
A heatmap indicating whether values are missing (0) or not (1). The plot can be retrieved from within R via \code{ComplexHeatmap::draw(mSetObj$imgSet$missval_heatmap.plot)}
(generated by \code{\link[ComplexHeatmap]{Heatmap}}).
}
\description{
\code{met.plot_missval} generates a heatmap of compounds
with missing values to discover whether values are missing at random or not.
}
