% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_PCA2DScore}
\alias{met.plot_PCA2DScore}
\title{Create 2D PCA score plot}
\usage{
met.plot_PCA2DScore(
  mSetObj = NA,
  imgName = "PCA_2DScores",
  format = "pdf",
  dpi = NULL,
  subtitle = FALSE,
  width = NA,
  pcx,
  pcy,
  reg = 0.95,
  show = 1,
  grey.scale = 0,
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container after principal component analysis (\code{\link[VisomX]{met.PCA.Anal}}).}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 dpi for PNG, 72 dpi for PDF)}

\item{subtitle}{(Logical, \code{TRUE} or \code{FALSE}) Shall the applied data transformation and scaling methods be displayed below the plot title?}

\item{width}{(Numeric) width of the the image file in inches (if \code{export = TRUE}).}

\item{pcx}{Specify the principal component on the x-axis}

\item{pcy}{Specify the principal component on the y-axis}

\item{reg}{Numeric, input a number between 0 and 1, 0.95 will display the 95 percent confidence regions, and 0 will not.}

\item{show}{Display sample names, \code{1} = show names, \code{0} = do not show names.}

\item{grey.scale}{Use grey-scale colors, \code{1} = grey-scale, \code{0} = not grey-scale.}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}
}
\value{
The input mSet object with added scatter plot. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$pca.score2d_PCx_PCy.plot)}.
}
\description{
\code{met.plot_PCA2DScore} visualizes clusters of samples based on their similarity in principal component analysis.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PlotPCA2DScore}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
}
