% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_PCAScree}
\alias{met.plot_PCAScree}
\title{Plot PCA scree plot}
\usage{
met.plot_PCAScree(
  mSetObj = NA,
  imgName = "PCA_ScreePlot",
  format = "pdf",
  dpi = NULL,
  width = NA,
  scree.num,
  plot = TRUE,
  export = FALSE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object.
Data container after principal component analysis (\code{\link[VisomX]{met.PCA.Anal}}).}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 dpi for PNG, 72 dpi for PDF)}

\item{width}{(Numeric) width of the the image file in inches (if \code{export = TRUE}).}

\item{scree.num}{Numeric, input a number to indicate the number of principal components to display in the scree plot.}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}
}
\value{
The input mSet object with added line plot. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$pca.scree.plot)}.
}
\description{
\code{met.plot_PCAScree} visualizes the proportion of variance explained for each principal component.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PlotPCAScree}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
}
