% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_PLS.Crossvalidation}
\alias{met.plot_PLS.Crossvalidation}
\title{Plot PLS-DA classification performance using different components}
\usage{
met.plot_PLS.Crossvalidation(
  mSetObj = NA,
  imgName = "PLSDA-CrossValidation",
  format = "pdf",
  dpi = NULL,
  width = NA,
  plot = TRUE,
  export = TRUE,
  title = FALSE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object.
Data container after partial least squares-discriminant analysis (\code{\link[VisomX]{met.PLSR.Anal}} and \code{\link[VisomX]{met.PLSDA.CV}}).}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 dpi for PNG, 72 dpi for PDF)}

\item{width}{(Numeric) width of the the image file in inches (if \code{export = TRUE}).}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}

\item{title}{(Logical) \code{TRUE} to add a title with the used normalization, transformation, and scaling method, or \code{FALSE} to not add any title.}
}
\value{
The input mSet object with added cross validation test plot. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$pls.crossvalidation.plot)}.
}
\description{
\code{met.plot_PLS.Crossvalidation} performs cross-validation (CV) on the generated PLS-DA model, where a fraction of data is held back, and the model trained on the rest. In each CV, the predicted data are compared with the original data, and the sum of squared errors is calculated. The prediction error is then summed over all samples (Predicted Residual Sum of Squares or PRESS). For convenience, the PRESS is divided by the initial sum of squares and subtracted from 1 to resemble the scale of the R2. Good predictions will have low PRESS or high Q2. Generally speaking, a model with an R2 (and Q2) value above 0.7 can be considered predictive. It is possible to have negative Q2, which means that your model is not at all predictive or is overfitted.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PlotPLS.Classification}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
