% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_PLS2DLoading}
\alias{met.plot_PLS2DLoading}
\title{Plot PLS-DA loadings and also set up the matrix for display}
\usage{
met.plot_PLS2DLoading(
  mSetObj = NA,
  imgName = "PLSDA_2DLoadings",
  format = "pdf",
  dpi = NULL,
  subtitle = FALSE,
  width = NA,
  inx1,
  inx2,
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container after partial least squares-discriminant analysis (\code{\link[VisomX]{met.PLSR.Anal}} and \code{\link[VisomX]{met.PLSDA.CV}}).}

\item{imgName}{(Character) Input a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 for PNG, 72 for PDF)}

\item{subtitle}{(\code{TRUE} or \code{FALSE}) Shall the applied data transformation and scaling methods be displayed below the plot title?}

\item{width}{(Numeric) width of the the image file in inch (if \code{export = TRUE})?}

\item{inx1}{Numeric, indicate the number of the principal component for the x-axis of the loading plot.}

\item{inx2}{Numeric, indicate the number of the principal component for the y-axis of the loading plot.}

\item{plot}{(\code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(\code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}
}
\value{
A scatter plot. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$mSetObj$pls.loading_PCx_PCy.plot)}.
}
\description{
\code{met.plot_PLS2DLoading} visualizes which features are responsible for the patterns seen among the samples in partial least squares-discriminant analysis.
}
