% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_PLS_Imp}
\alias{met.plot_PLS_Imp}
\title{Plot PLS important features}
\usage{
met.plot_PLS_Imp(
  mSetObj = NA,
  imgName,
  format = "pdf",
  dpi = NULL,
  width = NA,
  type = "vip",
  feat.nm = c("Comp. 1", "Comp. 2"),
  feat.num = 15,
  color.BW = FALSE,
  plot = TRUE,
  export = TRUE,
  title = FALSE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container after partial least squares-discriminant analysis (\code{\link[VisomX]{met.PLSR.Anal}} and \code{\link[VisomX]{met.PLSDA.CV}}).}

\item{imgName}{(Character) Input a name for the image file (if \code{export = TRUE}).}

\item{format}{(\code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 for PNG, 72 for PDF)}

\item{width}{(Numeric) width of the the image file in inch (if \code{export = TRUE})?}

\item{type}{(Character) Indicate the variables of importance type to use, \code{"vip"} to use VIP scores, or \code{"coef"} for coefficients}

\item{feat.nm}{(Character) Indicate the name of the feature. If \code{type = "vip"}, choose \code{"Comp. 1"}, \code{"Comp. 2"}, etc., depending on the component for which the VIP scores should be shown. If \code{type = "coef"}, choose \code{"coef.mean"} for average coefficients, or the name of a sample group in your data set.}

\item{feat.num}{(Numeric) Indicate the number of features to show in the plot.}

\item{color.BW}{(Logical) \code{TRUE} to use black and white, or \code{FALSE} to not.}

\item{plot}{(\code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(\code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}

\item{title}{(Logical) \code{TRUE} to add a title with the used normalization, transformation, and scaling method, or \code{FALSE} to not add any title.}
}
\value{
A dot plot. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$mSetObj$pls.imp_\strong{\emph{type}}_\strong{\emph{feat.nm}}.plot)}.
}
\description{
\code{met.plot_PLS_Imp} visualizes the features' importance in the PLS-DA model.
}
