% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_PLSImpScatter}
\alias{met.plot_PLSImpScatter}
\title{Plot PLS important features}
\usage{
met.plot_PLSImpScatter(
  mSetObj,
  imgName = "PLS_ImpScatter",
  format = "png",
  dpi = 300,
  width = NA,
  feat.nm = "coef.mean",
  vip.nm = c("Comp. 1", "Comp. 2"),
  plot = TRUE,
  export = FALSE,
  vip.thresh = 1,
  show.title = FALSE,
  title
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container after partial least squares-discriminant analysis (\code{\link[VisomX]{met.PLSR.Anal}} and \code{\link[VisomX]{met.PLSDA.CV}}).}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}).}

\item{format}{(\code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 dpi for PNG, 72 dpi for PDF)}

\item{width}{(Numeric) width of the the image file in inches (if \code{export = TRUE}).}

\item{feat.nm}{(Character) Indicate the name of the feature. Choose \code{"coef.mean"} for average coefficients, or the name of a sample group in your dataset.}

\item{vip.nm}{(Character or a character vector) Choose \code{"Comp. 1"}, \code{"Comp. 2"}, etc., depending on the component for which the VIP scores should be shown. If more than one component are specified (the default is \code{c("Comp. 1", "Comp. 2")}), an average VIP score is calculated.}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}

\item{vip.thresh}{(Numeric) Draw a vertical line in the plot indicating a chosen VIP relevance threshold (the default is \code{1}).}

\item{show.title}{(Logical) \code{TRUE} to add a title to the plot, or \code{FALSE} to not.}

\item{title}{(Character) Define the title if \code{show.title = TRUE}}
}
\value{
The input mSet object with added scatter plot. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$pls.ImpScatter_plot_feat.nm_.plot)}.
}
\description{
\code{met.plot_PLSImpScatter} visualizes the features' importance in the PLS-DA model by comparing VIP scores and coefficients.
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
}
