% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_SampleNormSummary}
\alias{met.plot_SampleNormSummary}
\title{Visualize effect of normalization on sample density distributions.}
\usage{
met.plot_SampleNormSummary(
  mSetObj = NA,
  imgName = "SampleNormSummary",
  format = "png",
  dpi = NULL,
  width = NA,
  show_prenorm = TRUE,
  export = TRUE,
  plot = TRUE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container after normalization (\code{\link[VisomX]{met.normalize}}).}

\item{imgName}{(Character) Input a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 for PNG, 72 for PDF)}

\item{width}{(Numeric) width of the the image file in inch (if \code{export = TRUE})?}

\item{show_prenorm}{(\code{TRUE} or \code{FALSE}) Shall density distributions before normalization be displayed?}

\item{export}{(\code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}

\item{plot}{(\code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}
}
\value{
The top is a density plot and the bottom is a box plot. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$summary_norm_sample.plot)}.
}
\description{
\code{met.plot_SampleNormSummary} visualizes the density distributions of samples before and after normalization, transformation, and scaling.
}
