% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_volcano}
\alias{met.plot_volcano}
\title{Volcano Plot}
\usage{
met.plot_volcano(
  mSetObj = NA,
  grp1,
  grp2,
  test = "ttest",
  paired = FALSE,
  nonpar = FALSE,
  equal.var = TRUE,
  log2fc.thresh = 1,
  threshp = 0.05,
  pval.type = "fdr",
  imgName = NULL,
  format = "pdf",
  add_names = TRUE,
  label_size = 3,
  dpi = NULL,
  width = NA,
  plot = TRUE,
  export = TRUE,
  silent = FALSE,
  test_condition = FALSE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container after ANOVA analysis (\code{\link[VisomX]{met.ANOVA.Anal}}) if \code{test = "anova"}.}

\item{grp1}{(Character) Enter name of the first group for the contrast \code{grp1 vs. grp2}. If both group arguments are empty, the first two names in the list of groups are selected.}

\item{grp2}{(Character) Enter name of the second group for the contrast \code{grp1 vs. grp2}. If both group arguments are empty, the first two names in the list of groups are selected.}

\item{test}{(Character) Choose a statistical tests. For \code{test = "ttest"}, \code{met.plot_volcano} runs \code{\link[VisomX]{met.Ttests.Anal}} with the chosen test parameters.
For \code{test = "anova"}, \code{\link[VisomX]{met.ANOVA.Anal}} must have been applied previously on the mSetObj.}

\item{paired}{(Logical) Is the data paired (\code{TRUE}) or not (\code{FALSE}). Only applicable for \code{test = "ttest"}.}

\item{nonpar}{(Logical) Use a non-parametric test (\code{TRUE}) or not (\code{FALSE}). Only applicable for \code{test = "ttest"}.}

\item{equal.var}{(Logical) The two groups have equal variance (\code{TRUE}) or not (\code{FALSE}). Only applicable for \code{test = "ttest"}.}

\item{log2fc.thresh}{(Numeric) Enter a relevance threshold for log2 fold changes, highlighted in the plot by vertical lines and colored compounds}

\item{threshp}{(Numeric) Enter a significance threshold for features based on T-test or ANOVA test results, highlighted in the plot by an horizontal line and colored compounds}

\item{pval.type}{(Character) Display and apply significance threshold to \code{"raw"} p values or adjusted p values (\code{"fdr"}).}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}). If \code{NULL}, the name \code{Plots/Volcano_grp1_vs_grp2} is assigned.}

\item{format}{(\code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{add_names}{(Logical) Display labels of significant features (\code{TRUE}) or not (\code{FALSE}).}

\item{label_size}{(Numeric) Font size for feature labels (if \code{add_names = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 dpi for PNG, 72 dpi for PDF)}

\item{width}{(Numeric) width of the the image file in inches (if \code{export = TRUE}).}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}

\item{silent}{(Logical) Shall the results of \code{\link[VisomX]{met.Ttests.Anal}} be printed in the console (\code{TRUE}) or not (\code{FALSE})?}

\item{test_condition}{(Logical) Add a subtitle with the applied data transformation and scaling methods be displayed below the plot title (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
The input mSet object with added volcano plot (generated by \code{\link[ggplot2]{ggplot}}). The plot can be retrieved from within R via \code{print(mSetObj$imgSet$volcano$grp1_vs_grp2.plot)}.
}
\description{
\code{met.plot_volcano} generates a volcano plot for a specified contrast.
}
\references{
adapted from \code{\link[DEP]{plot_volcano}} (\url{https://bioconductor.org/packages/DEP/}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
}
