% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.PLSDA.Permut}
\alias{met.PLSDA.Permut}
\title{Perform PLS-DA permutation}
\usage{
met.PLSDA.Permut(mSetObj = NA, num = 500, type = "accu")
}
\arguments{
\item{mSetObj}{Enter name of the created mSet object.}

\item{num}{(Numeric) Enter the number of permutations to perform.}

\item{type}{(Character) Type of accuracy indicator. \code{"accu"} for prediction accuracy, \code{"sep"} for separation distance.}
}
\description{
\code{met.read_data} performs PLS-DA permutation using training classification accuracy as
indicator, for two or more groups
}
\references{
adapted from \code{\link[MetaboAnalystR]{PLSDA.Permut}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
