% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.PLSR.Anal}
\alias{met.PLSR.Anal}
\title{Partial least squares (PLS) analysis using \emph{oscorespls} (Orthogonal scores algorithm)}
\usage{
met.PLSR.Anal(mSetObj = NA, reg = FALSE, data = "all")
}
\arguments{
\item{mSetObj}{Enter name of the created mSet object (see \code{\link[VisomX]{met.read_data}}).}

\item{reg}{(Logical)}

\item{data}{(Character) Enter \code{"all"} to train the PLS-DA model on your whole (filtered and normalized) data set or \code{"anova"} to use a subset of features defined as significant based on ANOVA analysis.}
}
\value{
The input mSet object with the results of PLS-DA at mSetObj$analSet$plsr.
}
\description{
\code{met.PLSR.Anal} performs PLS analysis using \emph{oscorespls}.
}
