% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.PreparePrenormData}
\alias{met.PreparePrenormData}
\title{Prepare data for normalization}
\usage{
met.PreparePrenormData(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Enter name of the created mSet object (see \code{\link[VisomX]{met.read_data}}).}
}
\value{
The input mSet object with prenorm data table at mSetObj$dataSet$prenorm.
}
\description{
\code{met.PreparePrenormData} checks for previous data editing and filtering, and defines adds a 'prenormalization' data table to the mSet object for further analysis.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PreparePrenormData}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
