% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.print_PLS3DLoading}
\alias{met.print_PLS3DLoading}
\title{Plot a generated 3D PLS-DA loading plot in RStudio}
\usage{
met.print_PLS3DLoading(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object.
Data container after \code{\link[VisomX]{met.plot_PLS3DLoading}}.}
}
\value{
A 3D scatter plot. To close the graphics device, execute \code{rgl::close3d()}.
}
\description{
\code{met.print_PLS3DLoading} opens a new graphics device to show the generated 3D PLS-DA loading plot.
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
}
