% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.read_data}
\alias{met.read_data}
\title{Construct mSet data container, read metabolomics data, filter data, and impute missing values}
\usage{
met.read_data(
  data,
  data.type = "conc",
  anal.type = "stat",
  paired = FALSE,
  csvsep = ";",
  format = "rowu",
  lbl.type = "disc",
  filt.feat = c(""),
  filt.smpl = c(""),
  filt.grp = c(""),
  filt.method = "none",
  remain.num = NULL,
  qcFilter = "F",
  qc.rsd = 25,
  all.rsd = NULL,
  imp.method = "lod",
  export = TRUE
)
}
\arguments{
\item{data}{Enter name of an R dataframe object or the "path name" (in quotes) of the CSV/TSV/XLS/XLSX/TXT file to read.}

\item{data.type}{(Character) The type of data, either "list" (Compound lists), \code{"conc"} (Compound concentration data), \code{"specbin"} (Binned spectra data), \code{"pktable"} (Peak intensity table), \code{"nmrpeak"} (NMR peak lists), \code{"mspeak"} (MS peak lists), or \code{"msspec"} (MS spectra data).}

\item{anal.type}{(Character) Indicate the analysis module to be performed: \code{"stat", "pathora", "pathqea", "msetora", "msetssp", "msetqea", "ts", "cmpdmap", "smpmap",} or \code{"pathinteg"}.}

\item{paired}{(Logical) Indicate if the data is paired (\code{TRUE}) or not (\code{FALSE}).}

\item{csvsep}{(Character) Enter the separator used in the CSV file (only applicable if reading a ".csv" file).}

\item{format}{(Character) Specify if samples are paired and in rows (\code{"rowp"}), unpaired and in rows (\code{"rowu"}), in columns and paired (\code{"colp"}), or in columns and unpaired (\code{"colu"}).}

\item{lbl.type}{(Character) Specify the group label type, either categorical (\code{"disc"}) or continuous (\code{"cont"}).}

\item{filt.feat}{(Character Vector) Enter the names of features to remove from the dataset.}

\item{filt.smpl}{(Character Vector) Enter the names of samples to remove from the dataset.}

\item{filt.grp}{(Character Vector) Enter the names of groups to remove from the dataset.}

\item{filt.method}{(Character) Select an option for unspecific filtering based on the following ranking criteria:
\itemize{
\item \code{"none"} apply no unspecific filtering.
\item \code{"rsd"} filters features with low relative standard deviation across the dataset.
\item \code{"nrsd"} is the non-parametric relative standard deviation.
\item \code{"mean"} filters features with low mean intensity value across the dataset.
\item \code{"median"} filters features with low median intensity value across the dataset.
\item \code{"sd"} filters features with low absolute standard deviation across the dataset.
\item \code{"mad"} filters features with low median absolute deviation across the dataset.
\item \code{"iqr"} filters features with a low inter-quartile range across the dataset.
}}

\item{remain.num}{(Numerical) Enter the number of variables to keep in your dataset. If \code{NULL}, the following empirical rules are applied during data filtering with the methods specified in \code{filter = ""}:
\itemize{
\item \strong{Less than 250 variables:} 5\% will be filtered
\item \strong{250 - 500 variables:} 10\% will be filtered
\item \strong{500 - 1000 variables:} 25\% will be filtered
\item \strong{More than 1000 variables:} 40\% will be filtered
}}

\item{qcFilter}{(Logical) Filter the variables based on the relative standard deviation of features in quality control (QC) samples (\code{TRUE}), or not (\code{FALSE}). This filter can be applied in addition to other, unspecific filtering methods.}

\item{qc.rsd}{(Numeric) Define the relative standard deviation cut-off in \%. Variables with a RSD greater than this number in the QC samples will be removed from the dataset. It is only necessary to specify this argument if \code{qcFilter} is \code{TRUE}. Otherwise, it will be ignored.}

\item{all.rsd}{(Numeric or \code{NULL}) Apply a filter based on the in-group relative standard deviation (RSD, in \%) or not \code{NULL}. Therefore, the RSD of every feature is calculated for every group in the dataset. If the RSD of a variable in any group exceeds the indicated threshold, it is removed from the dataset. This filter can be applied in addition to other filtering methods and is especially useful to perform on data with technical replicates.}

\item{imp.method}{(Character) Select the option to replace missing variables:
\itemize{
\item \code{"lod"} replaces missing values with 1/5 of the minimum value for the respective variable.
\item \code{"rowmin"} replaces missing values with the half sample minimum.
\item \code{"colmin"} replaces missing values with the half feature minimum.
\item \code{"mean"} replaces missing values with the mean value of the respective feature column.
\item \code{"median"} replaces missing values with the median value of the respective feature column.
\item \code{"knn_var"} imputes missing values by finding the features in the training set “closest” to it and averages these nearby points to fill in the value.
\item \code{"knn_smp"} imputes missing values by finding the samples in the training set “closest” to it and averages these nearby points to fill in the value.
\item \code{"bpca"} applies Bayesian PCA to impute missing values.
\item \code{"ppca"} applies probabilistic PCA to impute missing values.
\item \code{"svdImpute"} applies singular value decomposition to impute missing values.
}}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the missing value detection plots be exported as PDF or PNG file?}
}
\value{
An mSet object with (built in ascending order):
\itemize{
\item original data at \code{mSetObj$dataSet$data_orig}.
\item data with manually filtered out features/samples/groups at \code{mSetObj$dataSet$edit}.
\item data with unspecifically filtered data at  \code{mSetObj$dataSet$filt}.
\item data with imputed missing values at \code{mSetObj$dataSet$data_proc}.
\item missing value heatmap at \code{mSetObj$imgSet$missval_heatmap.plot} (see \code{\link[VisomX]{met.plot_missval}}).
\item Density and CumSum plots of intensities of proteins with and without missing values at \code{mSetObj$imgSet$missval_density.plot} (see \code{\link[VisomX]{met.plot_detect}}).
}
}
\description{
\code{met.read_data} is a wrapper function that constructs an mSet object, adds data from a table file or R dataframe object, applies unspecific and user-defined data filters, and imputes missing values.
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
