% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.report}
\alias{met.report}
\title{Generate a markdown report for metabolomics data analysis}
\usage{
met.report(mSetObj, report.dir = NULL, ...)
}
\arguments{
\item{mSetObj}{Enter name of the created mSet object after a full analysis workflow (see \code{\link[VisomX]{met.workflow}}).}

\item{report.dir}{(Character) Enter the name or location of the folder in which the report files are generated. If \code{NULL} (the default), a new folder "Report_\emph{date-time}" is created in your working directory.}

\item{...}{Further arguments passed from other methods.}
}
\value{
In the specified folder, this function creates a PDF report, an HTML report, and an .RData file including the entire mSet object.
}
\description{
\code{met.report} generates a report of the analysis performed by the \code{met.workflow} wrapper function.
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
