% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.SanityCheck}
\alias{met.SanityCheck}
\title{Perform sanity check on metabolomics data}
\usage{
met.SanityCheck(mSetObj = NA)
}
\arguments{
\item{mSetObj}{Enter the name of the created mSetObj (see \code{\link[VisomX]{met.initialize}})}
}
\description{
\code{met.SanityCheckData} is used for data processing, and performs a basic sanity
check of the uploaded content, ensuring that the data is suitable for further analysis.
The function will return a message if the data has successfully passed the check
and is deemed suitable for further analysis. If it fails, the function will return a 0.
The function will perform the check directly onto the mSet$dataSet object, and must
be performed immediately after reading in data.
The sanity check function evaluates the accuracy of sample and class labels, data structure,
deals with non-numeric values, and removes columns that are constant across all samples (variance = 0).
This function is performed automatically as part of \code{\link[VisomX]{met.read_data}}
}
\references{
adapted from \code{\link[MetaboAnalystR]{SanityCheckData}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
