% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.test_normalization}
\alias{met.test_normalization}
\title{Test for optimal data processing conditions}
\usage{
met.test_normalization(
  mSetObj,
  test_conditions = NULL,
  ref = NULL,
  class_order = FALSE,
  alpha = 0.05,
  lfc = 2,
  posthoc_method = "tukey",
  nonpar = FALSE,
  pls.cv.method = "LOOCV",
  pls.cv.k = 10,
  dpi = 300,
  pls.data = "all",
  permut.num = 500,
  vip.thresh = 1,
  report = FALSE,
  report.dir = NULL,
  export = FALSE,
  export.format = "pdf",
  export.dir = "met.Test_Normalization"
)
}
\arguments{
\item{mSetObj}{Enter the name of the created mSetObj (see \code{\link[VisomX]{met.read_data}})}

\item{test_conditions}{(Character vector) Enter the combinations of rowNorm, transNorm,
and scaleNorm separated by "/" to be tested in the workflow. See
\code{\link[VisomX]{met.normalize}} for suitable options. If \code{NA}, the
normalization test workflow will automatically choose \code{c("NULL", "MeanCenter",
  "LogNorm", "LogNorm/AutoNorm", "CrNorm", "CrNorm/AutoNorm", "AutoNorm", "RangeNorm",
  "ParetoNorm")} as default conditions. \emph{Please note}: rowNorm option "SpecNorm" is not supported by this workflow.}

\item{ref}{(Character vector) Enter the name(s) of the reference sample(s) or the
reference feature(s) for \code{rowNorm = "GroupPQN"}, \code{"SamplePQN"}, or
\code{"CompNorm"} in their respective order in \code{test_conditions}. Add "NULL" if the \code{ref} argument is not applicable to the respective test_condition.}

\item{class_order}{(Logical, \code{TRUE} or \code{FALSE}) Class order matters (i.e. implying time points, disease severity, etc.)}

\item{alpha}{(Numeric) Enter significance threshold for adjusted p values (false discovery rate - FDR; for ANOVA with post-hoc analyses).}

\item{lfc}{(Numeric) Enter relevance threshold for log2 fold changes in pair-wise comparisons (for ANOVA with post-hoc analyses).}

\item{posthoc_method}{(Character) Enter the name of the ANOVA post-hoc test, \code{"fisher"} or \code{"tukey"}.}

\item{nonpar}{(Logical) Use a non-parametric ANOVA test (\code{TRUE}) or not (\code{FALSE}).}

\item{pls.cv.method}{(Character) Enter one of two methods for PLS-DA model (cross) validation:
\itemize{
\item \code{"LOOCV"} performs leave-one-out cross validation
\item \code{"CV"} performs k-fold cross validation.
}}

\item{pls.cv.k}{(Numeric) The number of (randomized) groups that the dataset is to be split into during cross validation if \code{methodName = "CV"}. This value must be equal to or smaller than the number of samples.}

\item{dpi}{(Numeric) Resolution of PNG images (default is 300 dpi).}

\item{pls.data}{(Character) Enter \code{"all"} to train the PLS(-DA) model on your whole (filtered and normalized) dataset or \code{"anova"} to use a subset of features defined as significant based on ANOVA analysis.}

\item{permut.num}{(Numeric) Number of permutations in PLS-DA permutation tests.}

\item{vip.thresh}{(Numeric) Enter a chosen relevance threshold for PLS-DA VIP scores.}

\item{report}{(Logical) Generate a report with results of this workflow (\code{TRUE}) or not (\code{FALSE}).}

\item{report.dir}{(Character) Enter the name or location of the folder in which the
report files are generated if \code{report = TRUE}. If \code{NULL} (the default), a
new folder "Report_\emph{date-time}" is created in your working directory.}

\item{export}{(Logical) Export generated plots as PDF or PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{export.format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{export.dir}{(Character) Enter the name or location of a folder in which all generated files are saved.}
}
\description{
\code{met.test_normalization} runs a metabolomics analysis workflow with
several normalization/transformation/scaling combinations in parallel. The workflow
includes univariate analysis (ANOVA) and multivariate analyses (PCA and PLS-DA) and
creates a report with various performance indicators that help with deciding for
optimal data preprocessing conditions.
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
