% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.Ttests.Anal}
\alias{met.Ttests.Anal}
\title{Perform T-test analysis}
\usage{
met.Ttests.Anal(
  mSetObj = NA,
  grp1,
  grp2,
  nonpar = FALSE,
  threshp = 0.05,
  paired = FALSE,
  equal.var = TRUE,
  pvalType = "fdr",
  all_results = TRUE,
  silent = FALSE
)
}
\arguments{
\item{mSetObj}{Enter the name of the created mSet object (see \code{\link[VisomX]{met.read_data}}).}

\item{grp1}{(Character) Enter name of the first group for the contrast \code{grp1 vs. grp2}. If both group arguments are empty, the first two names in the list of groups are selected.}

\item{grp2}{(Character) Enter name of the second group for the contrast \code{grp1 vs. grp2}. If both group arguments are empty, the first two names in the list of groups are selected.}

\item{nonpar}{(Logical) Use a non-parametric test (\code{TRUE}) or not (\code{FALSE}).}

\item{threshp}{(Numerical) Enter the adjusted p-value (FDR) cutoff.}

\item{paired}{(Logical) Is the data paired (\code{TRUE}) or not (\code{FALSE}).}

\item{equal.var}{(Logical) Is the group variance equal (\code{TRUE}) or not (\code{FALSE}).}

\item{pvalType}{(Character) p value type used to apply significances based on the chosen threshold \code{threshp}. Can be \code{"fdr"} for adjusted p values or \code{"raw"} for raw p values.}

\item{all_results}{(Logical) Create a CSV file with T-Test results for all compounds (\code{TRUE}) or not (\code{FALSE}).}

\item{silent}{(Logical) Suppress message about the number of significant features found in the console (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
The input mSet object with T-Test results added at mSetObj$analSet$tt$grp1_vs_grp2.
}
\description{
\code{met.Ttests.Anal} performs the Student's t-test on chosen sample groups. For large datasets (> 1000 variables), both the paired information and the group variance will be ignored, and the default parameters will be used for t-tests to save computational time. If you choose non-parametric tests (Wilcoxon rank-sum test), the group variance will be ignored.
}
\references{
adapted from \code{\link[MetaboAnalystR]{Ttests.Anal}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
