% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.UpdateData}
\alias{met.UpdateData}
\title{Remove defined samples and features from the metabolomics dataset.}
\usage{
met.UpdateData(
  mSetObj = NA,
  filt.feat = c(""),
  filt.smpl = c(""),
  filt.grp = c("")
)
}
\arguments{
\item{mSetObj}{Enter the name of the created mSet object (see \code{\link[VisomX]{met.read_data}}).}

\item{filt.feat}{(Character vector) Enter the names of features (compounds) to remove from the dataset.}

\item{filt.smpl}{(Character vector) Enter the names of samples to remove from the dataset.}

\item{filt.grp}{(Character vector) Enter the names of groups (conditions) to remove from the dataset.}
}
\description{
\code{met.UpdateData} updates the mSet object after removing features or samples. This step is included in the data preparation workflow \code{\link[VisomX]{met.read_data}}.
}
\references{
adapted from \code{\link[MetaboAnalystR]{UpdateData}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
