% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.workflow}
\alias{met.workflow}
\title{Perform metabolomics data analysis workflow}
\usage{
met.workflow(
  mSetObj,
  rowNorm = "NULL",
  transNorm = "NULL",
  scaleNorm = "NULL",
  ref = NULL,
  norm.vec = NULL,
  contrast.type = "all",
  control = NULL,
  contrast = NULL,
  class_order = FALSE,
  alpha = 0.05,
  lfc = 2,
  posthoc_method = "tukey",
  nonpar = FALSE,
  permut.num = 500,
  vip.thresh = 1,
  volcano.test = "anova",
  volcano.add_names = TRUE,
  volcano.label_size = 3,
  volcano.adjusted = FALSE,
  pls.data = "all",
  pls.cv.method = "LOOCV",
  pls.cv.k = 5,
  plot = FALSE,
  export = FALSE,
  export.format = "pdf",
  dpi = 300,
  report = FALSE,
  report.nm = NULL,
  export.dir = "met.ProcessedData"
)
}
\arguments{
\item{mSetObj}{Enter the name of the created mSet object (see \code{\link[VisomX]{met.read_data}}).}

\item{rowNorm}{(Character) Select the option for row-wise normalization (see \code{\link[VisomX]{met.normalize}} for options).}

\item{transNorm}{(Character) Select option to transform the data (see \code{\link[VisomX]{met.normalize}} for options).}

\item{scaleNorm}{(Character) Select option for scaling the data (see \code{\link[VisomX]{met.normalize}} for options).}

\item{ref}{(Character) Enter the name of the reference sample or the reference feature (if \code{rowNorm = "GroupPQN"}, \code{"SamplePQN"}, or \code{"CompNorm"}.}

\item{norm.vec}{(Numeric vector) Vector with sample-specific scaling factors. Only applicable for \code{rowNorm = "SpecNorm"}.}

\item{class_order}{(Logical, \code{TRUE} or \code{FALSE}) Class order matters (i.e. implying time points, disease severity, etc.).}

\item{alpha}{(Numeric) Enter significance threshold for adjusted p values (false discovery rate - FDR; for ANOVA with post-hoc analyses).}

\item{lfc}{(Numeric) Enter relevance threshold for log2 fold changes in pair-wise comparisons (for ANOVA with post-hoc analyses).}

\item{posthoc_method}{(Character) Enter the name of the ANOVA post-hoc test, \code{"fisher"} or \code{"tukey"}.}

\item{nonpar}{(Logical) Use a non-parametric ANOVA test (\code{TRUE}) or not (\code{FALSE}).}

\item{permut.num}{(Numeric) Number of permutations in PLS-DA permutation tests.}

\item{vip.thresh}{(Numeric) Enter a chosen relevance threshold for PLS-DA VIP scores.}

\item{volcano.test}{(Character) Enter the name of the statistical, univariate analysis to perform, either \code{"anova"} or \code{"ttest"}.}

\item{volcano.add_names}{(Logical) Display compound labels next to dots within the volcano plot (\code{TRUE}) or not (\code{FALSE}).}

\item{volcano.label_size}{(Numerical) Enter the font size of compound labels within the volcano plot (if \code{volcano.add_names = TRUE})}

\item{volcano.adjusted}{(Logical) Display adjusted p-values on y axis of the volcano plot (\code{TRUE}) or not (\code{FALSE})? (only applicable to \code{volcano.test = "ttest"})}

\item{pls.data}{(Character) Enter \code{"all"} to train the PLS(-DA) model on your whole (filtered and normalized) dataset or \code{"anova"} to use a subset of features defined as significant based on ANOVA analysis.}

\item{pls.cv.method}{(Character) Enter one of two methods for PLS-DA model (cross) validation:
\itemize{
\item \code{"LOOCV"} performs leave-one-out cross validation
\item \code{"CV"} performs k-fold cross validation.
}}

\item{pls.cv.k}{(Numeric) The number of (randomized) groups that the dataset is to be split into during cross validation if \code{methodName = "CV"}. This value must be equal to or smaller than the number of samples.}

\item{plot}{(Logical) Display the generated plots in the Plots pane in RStudio (\code{TRUE}) or not (\code{FALSE}).}

\item{export}{(Logical) Export generated plots as PDF or PNG files (\code{TRUE}) or not (\code{FALSE}).}

\item{export.format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) Resolution of PNG images (default is 300 dpi).}

\item{report}{(Logical) Generate a report with results of this workflow (\code{TRUE}) or not (\code{FALSE}).}

\item{report.nm}{(Character) Enter the name or location of the folder in which the report files are generated if \code{report = TRUE}. If \code{NULL} (the default), a new folder "Report_\emph{date-time}" is created in your working directory.}

\item{export.dir}{(Character) Enter the name of location of a folder in which all generated files are saved.}
}
\description{
\code{met.workflow} runs a metabolomics analysis workflow, including
univariate analysis (ANOVA) and multivariate analyses (PCA and PLS-DA) and creates a
report with various graphical results.
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
