% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{Perform.permutation}
\alias{Perform.permutation}
\title{Permutation}
\usage{
Perform.permutation(perm.num, fun)
}
\arguments{
\item{perm.num}{(Numeric) Enter the number of permutations to perform.}

\item{fun}{Place holder for function.}
}
\description{
Perform permutation
}
\references{
adapted from \code{\link[MetaboAnalystR]{Perform.permutation}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
