% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot coverage of a SummarizedExperiment}
\usage{
plot_coverage(se, plot = TRUE)
}
\arguments{
\item{se}{A SummarizedExperiment}

\item{plot}{A logical indicating whether to return the coverage or not}
}
\value{
A data frame (if plot = FALSE) or a ggplot object
}
\description{
Plot coverage of a SummarizedExperiment
}
