% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fluctuator.R
\name{png_to_gif}
\alias{png_to_gif}
\title{...}
\usage{
png_to_gif(png, ..., fps = 0.5, out = "Plots/animated_gif.gif")
}
\value{
The input mSet object with imputed data at mSetObj$dataSet$data_proc.
}
\description{
...
}
