% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.boxplot_intensity}
\alias{prot.boxplot_intensity}
\title{Boxplot Intensity}
\usage{
prot.boxplot_intensity(se, ..., plot = TRUE, export = TRUE)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{...}{Additional SummarizedExperiment objects.}

\item{plot}{Logical. Should the plot be printed?}

\item{export}{Logical. Should the plot be exported as PDF and PNG files?}
}
\value{
(Invisibly) a ggplot object.
}
\description{
Generate boxplot for the intensity of each sample in a
SummarizedExperiment object.
}
