% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{prot.enricher_custom}
\alias{prot.enricher_custom}
\title{Title}
\usage{
prot.enricher_custom(
  gene,
  pvalueCutoff,
  pAdjustMethod = "BH",
  universe = NULL,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  USER_DATA
)
}
\arguments{
\item{USER_DATA}{}
}
\description{
Title
}
