% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{prot.filter_missing}
\alias{prot.filter_missing}
\title{Filter proteins based on missing values}
\usage{
prot.filter_missing(
  se,
  type = c("complete", "condition", "fraction", NULL),
  thr = NULL,
  min = NULL
)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object, proteomics data parsed with \code{\link{prot.read_data}}.}

\item{type}{(Character string) "complete", "condition" or "fraction", Sets the type of filtering applied. "complete" will only keep proteins with valid values in all samples. "condition" will keep proteins that have a maximum of \code{thr} missing values in at least one condition. "fraction" will keep proteins that have a \code{min} fraction of valid values in all samples.}

\item{thr}{(Integer) Sets the threshold for the allowed number of missing values in at least one condition if \code{type = "condition"}. In other words: "keep proteins that have a maximum of 'thr' missing values in at least one condition."}

\item{min}{(Numeric) Sets the threshold for the minimum fraction of valid values allowed for any protein if \code{type = "fraction"}.}
}
\value{
A filtered SummarizedExperiment object.
}
\description{
\code{prot.filter_missing} filters a proteomics dataset based on missing values. Different types of filtering can be applied, which range from only keeping proteins without missing values to keeping proteins with a certain percent valid values in all samples or keeping proteins that are complete in at least one condition.
}
