% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{prot.get_kegg_pathways}
\alias{prot.get_kegg_pathways}
\title{Multiple set version of union}
\usage{
prot.get_kegg_pathways(organism)
}
\arguments{
\item{x}{A list}
}
\description{
Multiple set version of union
}
