% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{prot.get_pathway_genes}
\alias{prot.get_pathway_genes}
\title{Get genes from pathway table}
\usage{
prot.get_pathway_genes(
  pathway_name,
  pathway_table,
  colid_pathways,
  colid_genes,
  gene_sep = ", "
)
}
\arguments{
\item{pathway_name}{(Character) Name of the pathway}

\item{pathway_table}{(Data frame) Pathway table}

\item{colid_pathways}{(Character) Column name of pathways}

\item{colid_genes}{(Character) Column name of genes}

\item{gene_sep}{(Character) Separator for genes}
}
\value{
(Character) Genes in the pathway
}
\description{
Get genes from pathway table
}
