% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{prot.get_results}
\alias{prot.get_results}
\title{Get Results}
\usage{
prot.get_results(dep)
}
\arguments{
\item{dep}{A SummarizedExperiment object}
}
\value{
A data frame containing the results
}
\description{
Function to generate a table with results of a differential expression analysis
}
