% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{prot.impute}
\alias{prot.impute}
\title{Title}
\usage{
prot.impute(
  se,
  fun = c("bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "man", "min", "zero",
    "mixed", "nbavg", "SampMin"),
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
