% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{prot.pca}
\alias{prot.pca}
\title{Title}
\usage{
prot.pca(
  mat,
  metadata = NULL,
  center = TRUE,
  scale = FALSE,
  rank = NULL,
  removeVar = NULL,
  transposed = FALSE,
  BSPARAM = ExactParam()
)
}
\arguments{
\item{BSPARAM}{}
}
\description{
Title
}
