% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_density}
\alias{prot.plot_density}
\title{Generates density plots for a SummarizedExperiment object}
\usage{
prot.plot_density(
  se1,
  se2 = NULL,
  se3 = NULL,
  title1 = "",
  title2 = "",
  title3 = "",
  group = FALSE,
  basesize = 15,
  plot = TRUE,
  export = FALSE
)
}
\arguments{
\item{se1}{A SummarizedExperiment object with normalized data.}

\item{se2}{(optional) A second SummarizedExperiment object with normalized data.}

\item{se3}{(optional) A third SummarizedExperiment object with normalized data.}

\item{title1}{A title for the first density plot.}

\item{title2}{(optional) A title for the second density plot.}

\item{title3}{(optional) A title for the third density plot.}

\item{group}{A logical value indicating whether to group the plots by condition (TRUE) or not (FALSE). If TRUE, \code{\link{prot.plot_imputation}} is called.}

\item{basesize}{The base font size for the plots.}

\item{plot}{A logical value indicating whether to plot the plots (TRUE) or not (FALSE).}

\item{export}{A logical value indicating whether to export the plots (TRUE) or not (FALSE).}
}
\value{
(Invisibly) returns a ggplot object.
}
\description{
Generates density plots for a SummarizedExperiment object
}
