% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_detect}
\alias{prot.plot_detect}
\title{Plot detectability for a SummarizedExperiment}
\usage{
prot.plot_detect(se, basesize = 10, plot = TRUE, export = TRUE)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{basesize}{Base font size for the plots}

\item{plot}{Logical, whether to plot the figure}

\item{export}{Logical, whether to export the figure}
}
\value{
Plots of the density distributions and cumulative fraction of proteins with and without missing values
}
\description{
Plots the group-wise density of proteins with missing values
}
