% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_enrichment}
\alias{prot.plot_enrichment}
\title{Plot pathway enrichment}
\usage{
prot.plot_enrichment(
  enrichset,
  title = "Differentially enriched pathways",
  subtitle = "",
  plot = TRUE,
  export = FALSE,
  kegg = TRUE
)
}
\arguments{
\item{enrichset}{A data frame containing the enrichment results.}

\item{title}{The title of the plot.}

\item{subtitle}{The subtitle of the plot.}

\item{plot}{Logical; whether to return the plot or not.}

\item{export}{Logical; whether to export the plot as PDF an PNG files}

\item{kegg}{Logical; whether the enrichment results are from KEGG or not.}
}
\value{
A ggplot object.
}
\description{
Plots the enrichment of pathways in a given dataset.
}
\seealso{
\code{\link{pathway_enrich}}
}
