% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_heatmap_all}
\alias{prot.plot_heatmap_all}
\title{Plots a heatmap of all protein abundances}
\usage{
prot.plot_heatmap_all(
  se,
  show_all = TRUE,
  pal = "RdBu",
  kmeans = FALSE,
  k = 6,
  col_limit = NA,
  indicate = "condition",
  clustering_distance = "euclidean",
  show_row_names = FALSE,
  row_font_size = 6,
  col_font_size = 10,
  plot = TRUE,
  export = TRUE,
  scale = TRUE,
  show_row_dend = FALSE,
  ...
)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{show_all}{Logical. Show protein abundances of all conditions or only of those in the specified constrast(s). Default is TRUE}

\item{pal}{Character. Color palette to use. Default is "RdBu"}

\item{kmeans}{Logical. Perform kmeans clustering on protein abundances. Default is FALSE}

\item{k}{Numeric. Number of clusters to use in kmeans clustering. Default is 6}

\item{col_limit}{Numeric. Maximum intensity for the heatmap. Default is NA}

\item{indicate}{Character. Column name of the annotation to indicate in the heatmap. Default is "condition"}

\item{clustering_distance}{Character. Distance metric to use for clustering. Default is "euclidean"}

\item{show_row_names}{Logical. Show protein names on the left of the heatmap. Default is FALSE}

\item{row_font_size}{Numeric. Font size for protein names. Default is 6}

\item{col_font_size}{Numeric. Font size for condition names. Default is 10}

\item{plot}{Logical. Plot the heatmap or not. Default is TRUE}

\item{export}{Logical. Export the heatmap to a pdf and png file. Default is TRUE}

\item{scale}{Logical. Scale the data before plotting. Default is TRUE}

\item{show_row_dend}{Logical. Show row dendrogram. Default is FALSE}

\item{...}{Other parameters to pass to ComplexHeatmap::Heatmap}
}
\value{
A data frame with protein abundances
}
\description{
Plots a heatmap of all protein abundances
}
