% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_missval}
\alias{prot.plot_missval}
\title{Plot missing values in SummarizedExperiment}
\usage{
prot.plot_missval(se, plot = TRUE, export = TRUE, fontsize = 12)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{plot}{Logical. Whether to return the plot.}

\item{export}{Logical. Whether to export the heatmap in pdf and png format.}

\item{fontsize}{Numeric. Font size for the column names in the heatmap.}
}
\value{
A binary heatmap of missing values in the SummarizedExperiment.
}
\description{
This function takes a SummarizedExperiment object and creates a binary heatmap of missing values in the assay.
}
