% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_screeplot}
\alias{prot.plot_screeplot}
\title{SCREE plot

prot.plot_screeplot() creates a SCREE plot of the explained variation in the principal components of a 'princomp' object}
\usage{
prot.plot_screeplot(
  pcaobj,
  components,
  xlim = NULL,
  ylim = c(0, 100),
  xlab = "Principal component",
  xlabAngle = 90,
  xlabhjust = 0.5,
  xlabvjust = 0.5,
  ylab = "Explained variation (\%)",
  ylabAngle = 0,
  ylabhjust = 0.5,
  ylabvjust = 0.5,
  axisLabSize = 16,
  title = "SCREE plot",
  subtitle = "",
  caption = "",
  titleLabSize = 16,
  subtitleLabSize = 12,
  captionLabSize = 12,
  colBar = "dodgerblue",
  drawCumulativeSumLine = TRUE,
  colCumulativeSumLine = "red2",
  sizeCumulativeSumLine = 1.5,
  drawCumulativeSumPoints = TRUE,
  colCumulativeSumPoints = "red2",
  sizeCumulativeSumPoints = 2,
  hline = NULL,
  hlineType = "longdash",
  hlineCol = "black",
  hlineWidth = 0.4,
  vline = NULL,
  vlineType = "longdash",
  vlineCol = "black",
  vlineWidth = 0.4,
  gridlines.major = TRUE,
  gridlines.minor = TRUE,
  borderWidth = 0.8,
  borderColour = "black",
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{pcaobj}{An object of class 'princomp'}

\item{components}{A numeric vector of principal component numbers to include in the plot}

\item{xlim}{A numeric vector of length 2 specifying the x-axis limits}

\item{ylim}{A numeric vector of length 2 specifying the y-axis limits. Default is c(0, 100)}

\item{xlab}{A character string for x-axis label}

\item{xlabAngle}{Numeric value for x-axis label angle}

\item{xlabhjust}{Numeric value for x-axis label horizontal justification}

\item{xlabvjust}{Numeric value for x-axis label vertical justification}

\item{ylab}{A character string for y-axis label}

\item{ylabAngle}{Numeric value for y-axis label angle}

\item{ylabhjust}{Numeric value for y-axis label horizontal justification}

\item{ylabvjust}{Numeric value for y-axis label vertical justification}

\item{axisLabSize}{Numeric value for axis label size}

\item{title}{A character string for the plot title}

\item{subtitle}{A character string for the plot subtitle}

\item{caption}{A character string for the plot caption}

\item{titleLabSize}{Numeric value for title label size}

\item{subtitleLabSize}{Numeric value for subtitle label size}

\item{captionLabSize}{Numeric value for caption label size}

\item{colBar}{A character string for the color of the bar chart}

\item{drawCumulativeSumLine}{Logical value to indicate if the cumulative sum line should be plotted}

\item{colCumulativeSumLine}{A character string for the color of the cumulative sum line}

\item{sizeCumulativeSumLine}{Numeric value for the size of the cumulative sum line}

\item{drawCumulativeSumPoints}{Logical value to indicate if the cumulative sum points should be plotted}

\item{colCumulativeSumPoints}{A character string for the color of the cumulative sum points}

\item{sizeCumulativeSumPoints}{Numeric value for the size of the cumulative sum points}

\item{hline}{A numeric value for horizontal line}

\item{hlineType}{A character string for the type of horizontal line}

\item{hlineCol}{A character string for the color of the horizontal line}

\item{hlineWidth}{Numeric value for the width of the horizontal line}

\item{vline}{A numeric value for vertical line}

\item{vlineType}{A character string for the type of vertical line}

\item{vlineCol}{A character string for the color of the vertical line}

\item{vlineWidth}{Numeric value for the width of the vertical line}

\item{gridlines.major}{Logical value to indicate if major gridlines should be plotted}

\item{gridlines.minor}{Logical value to indicate if minor gridlines should be plotted}

\item{borderWidth}{Numeric value for the width of the border}

\item{borderColour}{A character string for the color of the border}

\item{plot}{Logical value to indicate if the plot should be plotted}

\item{export}{Logical value to indicate if the plot should be exported as PNG and PDF}
}
\value{
A SCREE plot
}
\description{
SCREE plot

prot.plot_screeplot() creates a SCREE plot of the explained variation in the principal components of a 'princomp' object
}
