% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_upset}
\alias{prot.plot_upset}
\title{Plot an UpSet enrichment plot}
\usage{
prot.plot_upset(
  enrichset,
  order.by = "freq",
  point.size = 3,
  line.size = 1,
  text.scale = c(2, 2, 2, 2, 2, 1.5),
  ...
)
}
\arguments{
\item{enrichset}{An enrichment set object}

\item{order.by}{A character string indicating how to order the sets}

\item{point.size}{Numeric indicating the size of the points}

\item{line.size}{Numeric indicating the size of the lines}

\item{text.scale}{Numeric indicating the size of the text}

\item{...}{Further arguments to pass to the upset function}
}
\value{
An UpSetR plot
}
\description{
Plot an UpSet chart from an enrichment set
}
\details{
For more information about upset plots, see https://upset.app/
}
\examples{
prot.plot_upset(enrichset)

}
