% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_volcano}
\alias{prot.plot_volcano}
\title{Plots a volcano plot for a given contrast}
\usage{
prot.plot_volcano(
  dep,
  contrast,
  label_size = 3,
  alpha = 0.05,
  lfc = 2,
  add_names = TRUE,
  adjusted = FALSE,
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{dep}{A SummarizedExperiment object}

\item{contrast}{The contrast of interest in the format "conditionA_vs_conditionB"}

\item{label_size}{The size of labels for the protein names}

\item{alpha}{The alpha value used to determine significance.}

\item{lfc}{The log2 fold change threshold.}

\item{add_names}{Logical, whether to add protein names to the plot}

\item{adjusted}{Logical, whether to plot adjusted p-values. The \code{alpha} threshold is applied to adjusted p values even if \code{adjusted = FALSE}.}

\item{plot}{Logical, whether to return the volcano plot}

\item{export}{Logical, whether to export the volcano plot as PND and PDF file}
}
\value{
(invisibly) A data frame with columns for protein, log2 fold change, and -log10 p-value
}
\description{
Plots a volcano plot for a given contrast
}
