% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prot_workflow.R
\name{prot.workflow}
\alias{prot.workflow}
\title{Run a complete proteomics analysis workflow.}
\usage{
prot.workflow(
  se,
  normalize = TRUE,
  imp_fun = c("SampMin", "man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb",
    "min", "zero", "mixed", "nbavg"),
  q = 0.01,
  knn.rowmax = 0.5,
  type = c("all", "control", "manual"),
  control = NULL,
  contrast = NULL,
  alpha = 0.05,
  alpha_pathways = 0.1,
  lfc = 1,
  heatmap.show_all = TRUE,
  heatmap.kmeans = F,
  k = 6,
  heatmap.col_limit = NA,
  heatmap.show_row_names = TRUE,
  heatmap.row_font_size = 6,
  volcano.add_names = FALSE,
  volcano.label_size = 2.5,
  volcano.adjusted = TRUE,
  plot = FALSE,
  plot_volcano = FALSE,
  export = FALSE,
  report = TRUE,
  report.dir = NULL,
  pathway_enrichment = FALSE,
  pathway_kegg = FALSE,
  kegg_organism = NULL,
  custom_pathways = NULL,
  gsea = FALSE,
  gsea_gmt = NULL,
  gsea_pAdjustMethod = c("BH", "none"),
  gsea_pvalueCutoff = 0.05,
  gsea_minGSSize = 5,
  gsea_maxGSSize = 500,
  gsea_nPerm = 1000,
  out.dir = NULL
)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object, proteomics data parsed with \code{\link{prot.read_data}}.}

\item{normalize}{(Logical) Should the data be normalized via variance stabilization normalization?}

\item{imp_fun}{(Character string)  Function used for data imputation. "SampMin", "man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "min", "zero", "mixed", or "nbavg". See (\code{\link{prot.impute}}) for details.}

\item{q}{(Numeric) q value for imputing missing values with method \code{imp_fun = 'MinProb'}.}

\item{knn.rowmax}{(Numeric) The maximum percent missing data allowed in any row for \code{imp_fun = 'knn'}. Default: 0.5.}

\item{type}{(Character string) Type of differential analysis to perform. "all" (contrast each condition with every other condition), "control" (contrast each condition to a defined control condition), "manual" (manually define selected conditions).}

\item{control}{(Character string) The name of the control condition if \code{type = control}.}

\item{contrast}{(Character string or vector of strings) Define the contrasts to be tested if \code{type = manual} in the form: \code{"ConditionA_vs_ConditionB"}, or \code{c("ConditionA_vs_ConditionC", "ConditionB_vs_ConditionC")}.}

\item{alpha}{(Numeric) Significance threshold for adjusted p values.}

\item{alpha_pathways}{(Numeric) Significance threshold for adjusted p values in pathway enrichment analysis.}

\item{lfc}{(Numeric) Relevance threshold for log2(fold change) values. Only proteins with a |log2(fold change)| value above \code{lfc} for a given contrast are considered "significant" (if they additionally fullfil the \code{alpha} criterion).}

\item{heatmap.show_all}{(Logical) Shall all samples be displayed in the heat map (\code{TRUE}) or only the samples contained in the defined \code{contrast} (\code{FALSE})?}

\item{heatmap.kmeans}{(Logical) Shall the proteins be clustered in the heat map with the k-nearest neighbour method (\code{TRUE}) or not \code{FALSE})?}

\item{k}{(Integer) Number of protein clusters in heat map if \code{heatmap.kmeans = TRUE}.}

\item{heatmap.col_limit}{(Integer) Define the outer breaks in the heat map legend. Example: if \code{heatmap.col_limit = 3}, the color scale will span from -3 to 3. Alls values below -3 will have the same color as -3, and all values above 3 will have the same color as 3.}

\item{heatmap.show_row_names}{(Logical) Show protein names in heat map (\code{TRUE}) or not \code{FALSE}).}

\item{heatmap.row_font_size}{(Numeric) Font size of protein names in heat maps if \code{heatmap.show_row_names = TRUE}.}

\item{volcano.add_names}{(Logical) Show protein names in volcano plots (\code{TRUE}) or not \code{FALSE}).}

\item{volcano.label_size}{(Numeric) Font size of protein names in volcano plots if \code{volcano.add_names = TRUE}.}

\item{volcano.adjusted}{(Logical) Shall adjusted p values be shown on the y axis of volcano plots (\code{TRUE}) or raw p values (\code{FALSE})?.}

\item{plot}{(Logical) Show the generated plots in the \code{Plots} pane of RStudio (\code{TRUE}) or not \code{FALSE}).}

\item{plot_volcano}{(Logical) Generate volcano plots for each defined contrast (\code{TRUE}) or not \code{FALSE}).}

\item{export}{(Logical) Exported the generated plots as PNG and PDF files (\code{TRUE}) or not \code{FALSE}).}

\item{report}{(Logical) Render and export a report in PDF and HTML format that summarizes the results (\code{TRUE}) or not \code{FALSE}).}

\item{report.dir}{(Character string) Provide the name of or path to a folder into which the report will be saved.}

\item{pathway_enrichment}{(Logical) Perform pathway over-representation analysis for each tested contrast (\code{TRUE}) or not \code{FALSE}).}

\item{pathway_kegg}{(Logical) Perform pathway over-representation analysis with gene sets in the KEGG database (\code{TRUE}) or not \code{FALSE}).}

\item{kegg_organism}{(Character string) Identifier of the organism in the KEGG database (if \code{pathway_kegg = TRUE})}

\item{custom_pathways}{(a R dataframe object) Data frame providing custom pathway annotations. The table must contain a "Pathway" column listing identified pathway in the studies organism, and an "Accession" column listing the proteins (or genes) each pathway is composed of. The \strong{Accession} entries must match with protein \strong{IDs}.}

\item{out.dir}{(Character string) absolute path to the location where result TXT files should be exported to.}
}
\value{
A list containing \code{SummarizedExperiment} object for every computation step of the workflow, a \code{pca} object, and lists of up- or down regulated pathways for each tested contrast and method (KEGG and/or custom).
}
\description{
\code{prot.workflow} performs variance stabilization normalization (\code{\link{prot.normalize_vsn}}), missing value imputation (\code{\link{prot.impute}}), principal component analysis (\code{\link{prot.pca}}), differential enrichment test (\code{\link{prot.test_diff}}), and pathway enrichment analysis. If desired, standardized plots and a report are generated and exported as separate files.
}
