% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_visualization.R
\name{prot_to_map}
\alias{prot_to_map}
\title{Title}
\usage{
prot_to_map(
  se,
  protein_set = NULL,
  col.id = "ID",
  select.id = "Accession",
  pfx = NULL,
  imp_fun = c("man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "min", "zero",
    "mixed", "nbavg", "SampMin"),
  q = 0.01,
  type = c("contrast", "centered"),
  condition = NULL,
  contrast = NULL,
  template = NULL,
  RSD_thresh = 0.5,
  p_thresh = 1,
  indicate_nonsig = c("gray", "dashed"),
  result = NULL,
  pal = "RdYlBu",
  legend_min = NULL,
  legend_max = NULL,
  title = "",
  export = TRUE,
  inkscape_path = "C:/Program Files/Inkscape/bin/inkscape.exe",
  export_dpi = 300,
  export_width = 2281,
  export_height = 2166
)
}
\arguments{
\item{se}{\code{SummarizedExperiment} object, proteomics data parsed with \code{\link{prot.read_data}}.}

\item{protein_set}{Data frame containing a list of proteins to be plotted. The provided names in column \code{} must match the entries in the column provided as \code{col.id} as well as the IDs of objects in the map template.}

\item{col.id}{(Character) Column name of the protein identifier in the SummarizedExperiment object.}

\item{select.id}{Column name of the protein identifier in the \code{protein_set} data frame.}

\item{pfx}{Prefix of the protein identifiers to be plotted. all protein objects in the template map should have a common prefix (e.g., 'prot_')}

\item{imp_fun}{(Character string)  Function used for data imputation. "SampMin", "man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "min", "zero", "mixed", or "nbavg". See (\code{\link{prot.impute}}) for details.}

\item{q}{(Numeric) q value for imputing missing values with method \code{imp_fun = 'MinProb'}.}

\item{type}{(Character string) Type of analysis to perform. "contrast" (provide a specific contrast in the \code{contrast} argument to plot log2-fold change values) or "centered" (provide a single condition as \code{condition} to plot centered abundances in that condition).}

\item{condition}{Condition to be displayed if type = "centered".}

\item{contrast}{Tested contrast if type = "contrast" in the format "A_vs_B".}

\item{template}{Path to the SVG map template.}

\item{p_thresh}{(adjusted) p-value threshold.}

\item{result}{Name of the output files.}

\item{pal}{Brewer color palette used for protein squares.}

\item{legend_min}{Lower limit of the color legend.}

\item{legend_max}{Upper limit of the color legend.}

\item{title}{(Character string) Title shown above the map}

\item{export}{(Logical) A logical value indicating whether to export the plot as a PNG and PDF file.}

\item{inkscape_path}{The local path to the 'inkscape.exe' file. Required if \code{export = TRUE}.}

\item{export_dpi}{The dpi of the exported PNG and PDF file.}

\item{export_width}{The width of the exported PNG and PDF file. The width of the template SVG can be inspected after opening it in InkScape.}

\item{export_height}{The height of the exported PNG and PDF file. The height of the template SVG can be inspected after opening it in InkScape.}

\item{RSD_threshRelative}{standard deviation threshold.}
}
\value{
A metabolic map with protein objects colored according to their abundance or log2 fold changes in a given contrast.
}
\description{
Title
}
