% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fluctuator.R
\name{read_flux}
\alias{read_flux}
\title{Title}
\usage{
read_flux(
  file.data = dat_flux,
  csvsep = ";",
  dec = ".",
  sheet.data = 1,
  sheet.match = 1,
  rename.reaction = TRUE,
  rescale.reaction = NULL,
  file.match = "eq_bigg.csv",
  col.id = "Reaction ID",
  col.eq = "Equation",
  col.flux = "Value (\%)",
  col.sd = "Standard Deviation",
  FBA = FALSE
)
}
\arguments{
\item{FBA}{}
}
\description{
Title
}
