% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_computation.R
\name{rna.impute}
\alias{rna.impute}
\title{Impute missing values in a SummarizedExperiment object}
\usage{
rna.impute(
  se,
  fun = c("zero", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "man", "min",
    "zero", "mixed", "nbavg", "SampMin"),
  ...
)
}
\arguments{
\item{se}{A SummarizedExperiment object.}

\item{fun}{A character vector specifying the imputation method.
Options include "zero", "bpca", "knn", "QRILC", "MLE", "MinDet",
"MinProb", "man", "min", "zero", "mixed", "nbavg", and "SampMin".}

\item{...}{Additional arguments passed to the imputation method.}
}
\value{
A SummarizedExperiment object with missing values imputed.
}
\description{
This function imputes missing values in a SummarizedExperiment object,
using the specified imputation method.
}
\details{
"SampMin" replaces missing values with the minimum value found in each sample. For information about the remaining imputation methods, see \code{help("imputeMethods", "MsCoreUtils")}
}
