% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_computation.R
\name{rna.make_se}
\alias{rna.make_se}
\title{Create a SummarizedExperiment object}
\usage{
rna.make_se(genes_unique, columns, expdesign)
}
\arguments{
\item{genes_unique}{A data frame with unique gene names and columns of expression data.}

\item{columns}{A vector of the column positions or names with expression data in \code{genes_unique}.}

\item{expdesign}{A data frame of experimental design.}
}
\value{
A SummarizedExperiment object.
}
\description{
This function takes gene expression data stored in a data frame with unique gene names
and columns of expression data, as well as an experimental design data frame, and
creates a SummarizedExperiment object.
}
\details{
used internally by \code{rna.read_data()}.
}
