% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_plots.R
\name{rna.plot_bar}
\alias{rna.plot_bar}
\title{Plot bar plots for gene abundance or fold change}
\usage{
rna.plot_bar(
  dds,
  genes,
  type = c("contrast", "centered"),
  shrunken.lfc = T,
  contrast = NULL,
  col.id = "ID",
  name_table = NULL,
  match.id = "Accession",
  match.name = "Name",
  convert_name = FALSE,
  shape.size = 2.5,
  collage = TRUE,
  plot = TRUE,
  export = TRUE,
  export.nm = NULL
)
}
\arguments{
\item{dds}{A DESeqDataSet object containing the data.}

\item{genes}{A vector of gene names or IDs.}

\item{type}{Type of plot: "centered" or "contrast".}

\item{shrunken.lfc}{A logical indicating if shrunken log2 fold changes should be used if \code{type = "contrast"}. Default is TRUE.}

\item{contrast}{Name of the contrast for plotting when type is "contrast".}

\item{col.id}{Name of the column with gene IDs in \code{dep}.}

\item{name_table}{A data frame containing the mapping of gene names to reference names.}

\item{match.id}{The name of a column in \code{name_table} to find matches with \code{col.id}.}

\item{match.name}{The name of a column in \code{name_table} to assign new names based on matches of \code{col.id} and \code{match.id}.}

\item{convert_name}{Logical value indicating whether to convert the gene names to reference names based on \code{name_table} (e.g., gene names).}

\item{shape.size}{Numeric value defining the (replicate) symbol size in the plot}

\item{collage}{A logical indicating if multiple plots for different genes should be combined into collages of up to 8 plots. Default is TRUE.}

\item{plot}{wether to return the plot or not.}

\item{export}{Logical; whether to export the plot as PDF an PNG files}

\item{export.nm}{Name of the output PDF and PNG files if \code{export = TRUE}.}
}
\value{
If \code{plot} is set to FALSE and the length of \code{genes} is less than or equal to 8, a data frame with the log2 fold changes or log2 intensities, along with their lower and upper bounds of 95\% confidence intervals, is returned.
}
\description{
This function generates bar plots for gene abundance or fold change, normalized by either a reference gene or centered to the mean. The user can also plot the contrast of two conditions.
}
