% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_plots.R
\name{rna.plot_corrheatmap}
\alias{rna.plot_corrheatmap}
\title{Plot a correlation heatmap}
\usage{
rna.plot_corrheatmap(
  dds,
  lower = 0,
  upper = 1,
  pal = "PRGn",
  pal_rev = FALSE,
  indicate = NULL,
  font_size = 12,
  plot = TRUE,
  ...
)
}
\arguments{
\item{lower}{Numeric. Lower limit of the color scale. Default is 0.}

\item{upper}{Numeric. Upper limit of the color scale. Default is 1.}

\item{pal}{Character. Brewer color palette to be used. Default is "PRGn". See \code{\link[RColorBrewer]{brewer.pal.info}} for options}

\item{pal_rev}{Logical. If TRUE, color palette is reversed. Default is FALSE.}

\item{indicate}{Character. Column name from colData to use for annotations. Default is NULL.}

\item{font_size}{Numeric. Font size for labels. Default is 12.}

\item{plot}{Logical. If FALSE, no plot is generated, only data frame is returned. Default is TRUE.}

\item{...}{Other arguments to be passed to \code{\link[ComplexHeatmap]{Heatmap}}.}

\item{dep}{A DESeqDataSet object.}

\item{significant}{Logical. If TRUE, only significant correlations are plotted. Default is TRUE.}
}
\value{
(Invisibly) a data frame with Pearson correlation values is returned.
}
\description{
This function plots a Pearson correlation heatmap for a given SummarizedExperiment object.
}
