% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_plots.R
\name{rna.plot_heatmap}
\alias{rna.plot_heatmap}
\title{Plots a heatmap of the differentially expressed genes}
\usage{
rna.plot_heatmap(
  dds,
  type = c("centered", "contrast"),
  contrast = NULL,
  show_all = TRUE,
  pal = "RdBu",
  kmeans = FALSE,
  k = 6,
  col_limit = NA,
  indicate = NULL,
  clustering_distance = c("euclidean", "maximum", "manhattan", "canberra", "binary",
    "minkowski", "pearson", "spearman", "kendall", "gower"),
  show_row_names = FALSE,
  row_font_size = 6,
  col_font_size = 10,
  plot = TRUE,
  export = TRUE,
  ...
)
}
\arguments{
\item{dds}{A DESeqDataSet object containing the normalized gene abundances and the associated metadata}

\item{type}{Type of heatmap to plot. Either "centered" or "contrast".}

\item{contrast}{(String or vector of strings) Analyze only significant genes contained in the specified contrast(s). Default is NULL. Contrasts must be given in the form "ConditionA_vs_ConditionB"}

\item{show_all}{Show gene abundances of all conditions or only of those in the specified constrast(s). Default is TRUE}

\item{pal}{Color palette to use. Default is "RdBu"}

\item{kmeans}{Perform k-means clustering on the rows. Default is FALSE}

\item{k}{Number of clusters to use when kmeans is TRUE. Default is 6}

\item{col_limit}{Specify a custom range for the color scale of the heatmap. Default is NA, in which case the minimum and maximum of the color scale are calculated as the 5\% and 95\% percentiles.}

\item{indicate}{Annotate the heatmap with the specified column from the metadata.}

\item{clustering_distance}{Distance metric used for clustering. Default is "euclidean".}

\item{show_row_names}{Show the gene names on the left side of the heatmap. Default is FALSE.}

\item{row_font_size}{Font size for the gene names. Default is 6}

\item{col_font_size}{Font size for the condition names. Default is 10}

\item{plot}{Plot the heatmap. Default is TRUE}

\item{export}{Export the heatmap as pdf and png. Default is TRUE}

\item{...}{Other parameters passed to the ComplexHeatmap::Heatmap function}
}
\value{
(Invisibly) returns a data frame with the genes and the normalized abundances
}
\description{
Plots a heatmap of the differentially expressed genes
}
