% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_plots.R
\name{rna.plot_imputation}
\alias{rna.plot_imputation}
\title{Plot imputation results}
\usage{
rna.plot_imputation(
  assay,
  ...,
  colData,
  plot = TRUE,
  basesize = 12,
  export = TRUE
)
}
\arguments{
\item{assay}{A matrix or array of imputation results. If working with a DESeqDataSet or SummarizedExperiment object, the array can be retrieved via \code{SummarizedExperiment::assay(object)}.}

\item{...}{Additional matrices or arrays. Can be input as named argument, for example: \code{"matrix2" = log(matrix2)}}

\item{colData}{A data frame of sample metadata.  If working with a DESeqDataSet or SummarizedExperiment object, the metadata can be retrieved via \code{SummarizedExperiment::colData(object)}.}

\item{plot}{A logical indicating whether to plot the results.}

\item{basesize}{Base font size for plotting.}

\item{export}{A logical indicating whether to export the plot as PDF and PNG file into the /Plots folder.}
}
\value{
(Invisibly) A ggplot object.
}
\description{
This function takes a matrix or array of imputation results and plots the distribution of intensities. It is recommended to log-transform the data before using this function.
}
