% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_plots.R
\name{rna.plot_numbers}
\alias{rna.plot_numbers}
\title{Plot the number of genes identified in a DESeqDataSet object}
\usage{
rna.plot_numbers(se, plot = TRUE, export = FALSE, basesize = 12)
}
\arguments{
\item{se}{A DESeqDataSet object}

\item{plot}{Logical, if TRUE a plot is returned}

\item{export}{Logical, if TRUE the plot is exported to the /Plots directory}
}
\value{
If plot = TRUE, returns a ggplot object. If plot = FALSE, returns a data frame with plotted values.
}
\description{
Plot the number of genes identified in a DESeqDataSet object
}
