% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_plots.R
\name{rna.plot_pca}
\alias{rna.plot_pca}
\title{Plot Results of Principal Component Analysis}
\usage{
rna.plot_pca(
  dds,
  x = 1,
  y = 2,
  indicate = c("condition", "replicate"),
  title = NULL,
  label = FALSE,
  n = ncol(t(SummarizedExperiment::assay(dds))),
  point_size = 4,
  label_size = 3,
  hline = 0,
  hlineType = "longdash",
  hlineCol = "black",
  hlineWidth = 0.4,
  vline = 0,
  vlineType = "longdash",
  vlineCol = "black",
  vlineWidth = 0.4,
  basesize = 15,
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{x}{x-axis PC (default: 1)}

\item{y}{y-axis PC (default: 2)}

\item{indicate}{features to indicate in the plot ("condition" or "replicate")}

\item{title}{title of the plot (default: "PCA plot - top \code{n} variable genes")}

\item{label}{whether to label points in the plot (default: FALSE)}

\item{n}{number of variables to take into account, sorted according to their variance in descending order. Only the \code{n} most variable genes are used to perform PCA. default: number of columns in \code{dep})}

\item{point_size}{size of points in the plot (default: 4)}

\item{label_size}{size of labels in the plot (default: 3)}

\item{hline}{position of horizontal line (default: 0)}

\item{hlineType}{type of horizontal line (default: 'longdash')}

\item{hlineCol}{color of horizontal line (default: 'black')}

\item{hlineWidth}{width of horizontal line (default: 0.4)}

\item{vline}{position of vertical line (default: 0)}

\item{vlineType}{type of vertical line (default: 'longdash')}

\item{vlineCol}{color of vertical line (default: 'black')}

\item{vlineWidth}{width of vertical line (default: 0.4)}

\item{basesize}{base size of the plot (default: 15)}

\item{plot}{whether to return the PCA plot (default: TRUE)}

\item{export}{whether to export the PCA plot to the Plots directory as PNG and PDF file (default: TRUE)}

\item{dep}{DESeqDataSet object}
}
\value{
(invisibly) a data frame containing the PCA coordinates
}
\description{
Performs 'regularized log' transformation followed by PCA on a given DESeqDataSet object and plots the results.
}
