% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_computation.R
\name{rna.report}
\alias{rna.report}
\title{Generate a transcriptomics analysis report}
\usage{
rna.report(results, report.dir = NULL, ...)
}
\arguments{
\item{results}{A list object containing the results of an RNA analysis workflow.}

\item{report.dir}{A character string specifying the directory in which to save the report.}

\item{...}{Additional parameters to be passed to the report.}
}
\value{
A report in the specified directory.
}
\description{
This function generates a report based on the results of an RNA analysis workflow.
}
\examples{

rna.report(results, report.dir = NULL)
}
