% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_misc_utils.R
\name{Setdiff}
\alias{Setdiff}
\title{Remove the union of the y's from the common x's.}
\usage{
Setdiff(x, y)
}
\arguments{
\item{x}{A list of characters}

\item{y}{A list of characters}
}
\description{
Remove the union of the y's from the common x's.
}
