% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{theme_DEP1}
\alias{theme_DEP1}
\title{theme_DEP1}
\usage{
theme_DEP1(basesize = 15)
}
\arguments{
\item{basesize}{numeric. Base font size for the theme.}
}
\value{
A ggplot2 theme object.
}
\description{
A function to create a theme for ggplot2
}
