% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{theme_DEP2}
\alias{theme_DEP2}
\title{Theme DEP2}
\usage{
theme_DEP2(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link{theme_DEP1}}}
}
\value{
A modified theme
}
\description{
Modify the theme DEP1
}
