% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fluctuator.R
\name{flux_to_map}
\alias{flux_to_map}
\title{...}
\usage{
flux_to_map(
  df = "dat_flux.",
  template = NULL,
  result = NULL,
  pal = "G",
  title = "",
  export = TRUE,
  export_dpi = 300,
  export_width = 2281,
  export_height = 2166,
  FBA = FALSE
)
}
\arguments{
\item{df}{data frame}
}
\value{
The input mSet object with imputed data at mSetObj$dataSet$data_proc.
}
\description{
...
}
