% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{make_se}
\alias{make_se}
\title{Create a SummarizedExperiment object from a data frame}
\usage{
make_se(proteins_unique, columns, expdesign)
}
\arguments{
\item{proteins_unique}{A data frame containing the protein expression data}

\item{columns}{A vector of column numbers containing the numeric expression data}

\item{expdesign}{A data frame containing the experimental design}
}
\value{
A SummarizedExperiment object
}
\description{
This function creates a SummarizedExperiment object from a data frame containing protein expression data.
}
