% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_ANOVA}
\alias{met.plot_ANOVA}
\title{Visualize p values determined by ANOVA analysis.}
\usage{
met.plot_ANOVA(
  mSetObj = NA,
  imgName = "ANOVA_plot",
  format = "pdf",
  dpi = NULL,
  width = NA,
  subtitle = FALSE,
  export = TRUE,
  plot = TRUE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container after ANOVA analysis (\code{\link[VisomX]{met.ANOVA.Anal}}).}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 dpi for PNG, 72 dpi for PDF)}

\item{width}{(Numeric) width of the the image file in inches (if \code{export = TRUE}).}

\item{subtitle}{(Logical, \code{TRUE} or \code{FALSE}) Shall the applied data transformation and scaling methods be displayed below the plot title?}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}
}
\value{
The input mSet object with added scatter plot with the p value for each compound analyzed. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$anova.plot)}.
(generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{met.plot_ANOVA} visualizes the significance of each compound in the dataset.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PlotANOVA}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
}
