% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_CorrHeatMap_Samples}
\alias{met.plot_CorrHeatMap_Samples}
\title{Pattern hunter, plot correlation heatmap}
\usage{
met.plot_CorrHeatMap_Samples(
  mSetObj = NA,
  imgName = "correlation_samples",
  format = "pdf",
  dpi = 72,
  width = NA,
  cor.method = "pearson",
  colors = "bwm",
  viewOpt = "overview",
  fix.col = FALSE,
  no.clst = FALSE,
  corrCutoff = 0,
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object.}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 dpi for PNG, 72 dpi for PDF)}

\item{width}{(Numeric) width of the the image file in inches (if \code{export = TRUE}).}

\item{cor.method}{(Character) Indicate the correlation method, \code{'pearson'}, \code{'spearman'}, or \code{'kendall'}.}

\item{colors}{(Character) Indicate the colors for the heatmap, \code{"bwm"} for default, \code{"gbr"} for red/green, \code{"heat"} for heat colors, \code{"topo"} for topo colors, and \code{"gray"} for gray scale.}

\item{viewOpt}{(Character) Indicate \code{"overview"} to get an overview of the heatmap, and \code{"detail"} to get a detailed view of the heat map.}

\item{fix.col}{(Logical) Fix colors (\code{TRUE}) or not (\code{FALSE}).}

\item{no.clst}{(Logical) Indicate if the correlations should be clustered (\code{TRUE}) or not (\code{FALSE}).}

\item{corrCutoff}{(Numeric) Set correlation cut off.}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}
}
\value{
The input mSet object with added heat map showing the correlations between samples. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$corr.heatmap_samples.plot)}.
}
\description{
\code{met.plot_CorrHeatMap_Samples} visualizes the correlations between samples.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PlotCorrHeatMap}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
}
