% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_plots.R
\name{met.plot_PLS2DScore}
\alias{met.plot_PLS2DScore}
\title{Create 2D PLS-DA score plot}
\usage{
met.plot_PLS2DScore(
  mSetObj = NA,
  imgName = "PLSDA_2DScore",
  format = "pdf",
  dpi = NULL,
  width = NA,
  inx1 = 1,
  inx2 = 2,
  reg = 0.95,
  show = 1,
  grey.scale = 0,
  subtitle = FALSE,
  use.sparse = FALSE,
  plot = TRUE,
  export = TRUE
)
}
\arguments{
\item{mSetObj}{Input name of the created mSet object,
Data container after partial least squares-discriminant analysis (\code{\link[VisomX]{met.PLSR.Anal}} and \code{\link[VisomX]{met.PLSDA.CV}}).}

\item{imgName}{(Character) Enter a name for the image file (if \code{export = TRUE}).}

\item{format}{(Character, \code{"png"} or \code{"pdf"}) image file format (if \code{export = TRUE}).}

\item{dpi}{(Numeric) resolution of the image file (if \code{export = TRUE}). If \code{NULL}, the resolution will be chosen automatically based on the  chosen file format (300 dpi for PNG, 72 dpi for PDF)}

\item{width}{(Numeric) width of the the image file in inches (if \code{export = TRUE}).}

\item{inx1}{(Numeric) Indicate the number of the principal component for the x-axis of the loading plot.}

\item{inx2}{(Numeric) Indicate the number of the principal component for the y-axis of the loading plot.}

\item{reg}{(Numeric) Enter a number between 0 and 1, 0.95 will display the 95 percent confidence regions, and 0 will not.}

\item{show}{Display sample names, \code{1} = show names, \code{0} = do not show names.}

\item{grey.scale}{Use grey-scale colors, \code{1} = grey-scale, \code{0} = not grey-scale.}

\item{subtitle}{(Logical, \code{TRUE} or \code{FALSE}) Shall the applied data transformation and scaling methods be displayed below the plot title?}

\item{use.sparse}{(Logical) Use a sparse algorithm (\code{TRUE}) or not (\code{FALSE}).}

\item{plot}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be returned in the RStudio 'Plots' pane?}

\item{export}{(Logical, \code{TRUE} or \code{FALSE}) Shall the plot be exported as PDF or PNG file?}
}
\value{
The input mSet object with added scatter plot. The plot can be retrieved from within R via \code{print(mSetObj$imgSet$pls.score2d_PCx_PCy.plot)}.
}
\description{
\code{met.plot_PLS2DScore} visualizes clusters of samples based on their similarity in partial least squares-discriminant analysis.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PlotPLS2DScore}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
\author{
Nicolas T. Wirth \email{mail.nicowirth@gmail.com}
Technical University of Denmark
License: GNU GPL (>= 2)
}
