% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolomics_computation.R
\name{met.PLSDA.CV}
\alias{met.PLSDA.CV}
\title{PLS-DA classification and feature selection}
\usage{
met.PLSDA.CV(
  mSetObj = NA,
  methodName = "CV",
  compNum = GetDefaultPLSCVComp(mSetObj),
  choice = "Q2",
  data = "all"
)
}
\arguments{
\item{mSetObj}{Enter name of the created mSet object after PLS model creation (see \code{\link[VisomX]{met.PLSR.Anal}}).}

\item{methodName}{(Character) Enter one of two methods for PLS-DA model (cross) validation:
\itemize{
\item \code{"LOOCV"} performs leave-one-out cross validation
\item \code{"CV"} performs 10-fold cross validation.
}}

\item{compNum}{Number of components used for model validation (detected automatically).}

\item{choice}{(Character) Choose the criterion used to estimate the predictive ability of the model, \code{"Q2"} or \code{"R2"}.}

\item{data}{(Character) Enter \code{"all"} to train the PLS-DA model on your whole (filtered and normalized) data set or \code{"anova"} to use a subset of features defined as significant based on ANOVA analysis.}
}
\value{
The input mSet object with the results of PLS-DA at mSetObj$analSet$plsda.
}
\description{
\code{met.PLSDA.CV} performs group classification and feature selection on a PLS model.
}
\references{
adapted from \code{\link[MetaboAnalystR]{PLSDA.CV}} (\url{https://github.com/xia-lab/MetaboAnalystR}).
}
