% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{pathway_enrich}
\alias{pathway_enrich}
\title{Title}
\usage{
pathway_enrich(
  gene,
  organism = "ppu",
  keyType = "kegg",
  pvalueCutoff = 0.05,
  pAdjustMethod = "BH",
  universe,
  minGSSize = 10,
  maxGSSize = 500,
  qvalueCutoff = 0.2,
  use_internal_kegg = FALSE,
  custom_gene_sets = FALSE,
  custom_pathways = NULL
)
}
\arguments{
\item{custom_pathways}{}
}
\description{
Title
}
