% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{prot.impute}
\alias{prot.impute}
\title{Impute missing values in a SummarizedExperiment object}
\usage{
prot.impute(
  se,
  fun = c("bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "man", "min", "zero",
    "mixed", "nbavg", "SampMin"),
  ...
)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{fun}{Character; imputation method, one of \code{"bpca"}, \code{"knn"}, \code{"QRILC"}, \code{"MLE"}, \code{"MinDet"}, \code{"MinProb"}, \code{"man"}, \code{"min"}, \code{"zero"}, \code{"mixed"}, \code{"nbavg"}, or \code{"SampMin"}.}

\item{...}{Additional arguments passed to the imputation function.}
}
\value{
A SummarizedExperiment object with imputed values
}
\description{
Imputes missing values in a SummarizedExperiment object.
}
\details{
"SampMin" replaces missing values with the minimum value found in each sample. For information about the remaining imputation methods, see \code{help("imputeMethods", "MsCoreUtils")}
}
\examples{
data(mset)
se <- make_se(mset)
se <- prot.impute(se, fun = "knn")

}
\seealso{
\code{\link{prot.make_unique}}, \code{\link{prot.make_se}}
}
\keyword{SummarizedExperiment}
\keyword{imputation,}
