% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_bar}
\alias{prot.plot_bar}
\title{Plot bar plots for protein abundance or fold change}
\usage{
prot.plot_bar(
  dep,
  proteins,
  combine = FALSE,
  ref.prot = NULL,
  type = c("contrast", "centered", "reference", "abundance"),
  contrast = NULL,
  col.id = "ID",
  match.id = "Accession",
  match.name = "Name",
  convert_name = FALSE,
  shape.size = 2.5,
  y.lim = NULL,
  name_table = NULL,
  plot = TRUE,
  export = TRUE,
  export.nm = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{dep}{A SummarizedExperiment object containing the data.}

\item{proteins}{A vector of protein names or IDs.}

\item{combine}{Logical value indicating whether to combine the data from different conditions in one plot.}

\item{ref.prot}{ID or name of the reference protein for normalization when type is "reference".}

\item{type}{Type of plot: "abundance", "reference", "centered" or "contrast".}

\item{contrast}{Name of the contrast for plotting when type is "contrast".}

\item{match.id}{A vector of protein names or IDs that need to be matched in the name.table.}

\item{match.name}{A vector of protein names that need to be matched in the name.table.}

\item{y.lim}{Limits of the y axis.}

\item{subset}{A data frame containing a subset of the data to be plotted.}

\item{convert.name}{Logical value indicating whether to convert the protein names to reference names (e.g., gene names).}

\item{name.table}{A data frame containing the mapping of protein names to reference names.}
}
\value{
A ggplot object as well as (invisibly) a dataframe with the data used for creating the plot.
}
\description{
This function generates bar plots for protein abundance or fold change, normalized by either a reference protein or centered to the mean. The user can also plot the contrast of two conditions.
}
