% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_bar}
\alias{prot.plot_bar}
\title{Plot bar plots for protein abundance or fold change}
\usage{
prot.plot_bar(
  dep,
  proteins,
  combine = FALSE,
  type = c("contrast", "centered", "ReferenceProt", "ReferenceCondition", "abundance"),
  ref.prot = NULL,
  ref.condition = NULL,
  contrast = NULL,
  col.id = "ID",
  name_table = NULL,
  match.id = "Accession",
  match.name = "Name",
  convert_name = FALSE,
  shape.size = 2.5,
  y.lim = NULL,
  plot = TRUE,
  export = FALSE,
  export.nm = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{dep}{A SummarizedExperiment object containing the data.}

\item{proteins}{A vector of protein names or IDs.}

\item{combine}{Logical value indicating whether to combine the data from different proteins in one plot.}

\item{type}{Type of plot: "abundance", "ReferenceProt", "centered" or "contrast".}

\item{ref.prot}{ID or name of the reference protein for normalization when type is "ReferenceProt".}

\item{ref.condition}{Name of the reference condition by which all average abundance values are divided when type is "ReferenceCondition".}

\item{contrast}{Name of the contrast for plotting when type is "contrast".}

\item{col.id}{Name of the column with protein IDs in \code{dep}.}

\item{name_table}{A data frame containing the mapping of protein names to reference names.}

\item{match.id}{The name of a column in \code{name_table} to find matches with \code{col.id}.}

\item{match.name}{The name of a column in \code{name_table} to assign new names based on matches of \code{col.id} and \code{match.id}.}

\item{convert_name}{Logical value indicating whether to convert the protein names to reference names based on \code{name_table} (e.g., gene names).}

\item{shape.size}{Numeric value defining the (replicate) symbol size in the plot}

\item{y.lim}{Limits of the y axis.}

\item{plot}{wether to return the plot or not.}

\item{export}{Logical; whether to export the plot as PDF an PNG files}

\item{export.nm}{Name of the output PDF and PNG files if \code{export = TRUE}.}

\item{width}{Numeric value defining the width of the exported plot.}

\item{height}{Numeric value defining the height of the exported plot.}
}
\value{
(invisibly) a ggplot object as well as the plot in the \link{Plots} pane if \code{plot = TRUE}.
}
\description{
This function generates bar plots for protein abundance or fold change, normalized by either a reference protein or centered to the mean. The user can also plot the contrast of two conditions.
}
