% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_plots.R
\name{prot.plot_numbers}
\alias{prot.plot_numbers}
\title{Plot the number of proteins identified in a SummarizedExperiment object}
\usage{
prot.plot_numbers(se, plot = TRUE, export = FALSE)
}
\arguments{
\item{se}{A SummarizedExperiment object}

\item{plot}{Logical, if TRUE a plot is returned}

\item{export}{Logical, if TRUE the plot is exported to the /Plots directory}
}
\value{
If plot = TRUE, returns a ggplot object. If plot = FALSE, returns a data frame.
}
\description{
Plot the number of proteins identified in a SummarizedExperiment object
}
