% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteomics_computation.R
\name{prot.test_diff}
\alias{prot.test_diff}
\title{Title}
\usage{
prot.test_diff(
  se,
  type = c("control", "all", "manual"),
  control = NULL,
  test = NULL,
  design_formula = stats::formula(~0 + condition)
)
}
\arguments{
\item{design_formula}{}
}
\description{
Title
}
