% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_fluctuator.R
\name{prot_to_map}
\alias{prot_to_map}
\title{Title}
\usage{
prot_to_map(
  se,
  protein_set = NULL,
  col.id = "ID",
  select.id = "Accession",
  pfx = NULL,
  imp_fun = c("man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "min", "zero",
    "mixed", "nbavg", "SampMin"),
  q = 0.01,
  type = c("contrast", "centered"),
  condition = NULL,
  contrast = NULL,
  template = NULL,
  RSD_thresh = 0.5,
  p_thresh = 1,
  indicate_nonsig = c("gray", "dashed"),
  result = NULL,
  pal = "RdYlBu",
  legend_min = NULL,
  legend_max = NULL,
  title = "",
  export = TRUE,
  export_dpi = 300,
  export_width = 2281,
  export_height = 2166
)
}
\arguments{
\item{export_height}{}
}
\description{
Title
}
