% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_plots.R
\name{rna.plot_bar}
\alias{rna.plot_bar}
\title{Plot bar plots for gene abundance or fold change}
\usage{
rna.plot_bar(
  dds,
  genes,
  type = c("contrast", "log_centered", "centered", "raw", "log2"),
  shrunken.lfc = TRUE,
  contrast = NULL,
  conditions = NULL,
  col.id = "ID",
  name_table = NULL,
  match.id = "Accession",
  match.name = "Name",
  convert_name = FALSE,
  shape.size = 2.5,
  collage = TRUE,
  plot = TRUE,
  export = TRUE,
  export.nm = NULL
)
}
\arguments{
\item{dds}{A DESeqDataSet object containing the data.}

\item{genes}{A vector of gene names or IDs.}

\item{type}{Type of plot: "contrast", "log_centered", "centered", "raw", or "log2".
\itemize{
\item "log_centered": log2-transformed intensities centered to their mean.
\item "centered": raw intensities centered to their mean.
\item "contrast": log2 fold changes for specified contrasts.
\item "raw": raw (or normalized) read counts with mean and 95\% CI.
\item "log2": log2-transformed intensities with mean and 95\% CI.
}}

\item{shrunken.lfc}{A logical indicating if shrunken log2 fold changes should be used if \code{type = "contrast"}. Default is TRUE.}

\item{contrast}{Name of the contrast for plotting when type is "contrast".}

\item{conditions}{Character vector of condition names to be plotted.}

\item{col.id}{Name of the column with gene IDs in \code{dds}.}

\item{name_table}{A data frame containing the mapping of gene names to reference names.}

\item{match.id}{The name of a column in \code{name_table} to find matches with \code{col.id}.}

\item{match.name}{The name of a column in \code{name_table} to assign new names based on matches of \code{col.id} and \code{match.id}.}

\item{convert_name}{Logical indicating whether to convert gene names based on \code{name_table}.}

\item{shape.size}{Numeric; replicate symbol size in the plot.}

\item{collage}{Logical; combine multiple gene plots into collages of up to 8. Default TRUE.}

\item{plot}{Logical; print the plot if TRUE or return data frame if FALSE. Default TRUE.}

\item{export}{Logical; export plots as PDF/PNG if TRUE. Default TRUE.}

\item{export.nm}{Base name for exported files if \code{export = TRUE}.}
}
\value{
If \code{plot = FALSE} and length(genes) <= 8, returns a data frame of values and 95\% CIs.
}
\description{
This function generates bar plots for gene abundance or fold change, normalized by either a reference gene or centered to the mean. The user can also plot log2 intensities, contrasts, or raw counts.
}
