% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_computation.R
\name{rna.read_data}
\alias{rna.read_data}
\title{Read Transcriptomics Data}
\usage{
rna.read_data(
  data = NULL,
  files.ind = NULL,
  expdesign = NULL,
  csvsep = ";",
  dec = ".",
  sheet = 1,
  name = "SymbolID",
  id = "gene_id",
  values = "FPKM",
  id2name.table = NULL,
  id2name.id = NULL,
  id2name.name = NULL,
  pfx.counts = "counts.",
  rsd_thresh = NULL,
  filt_type = c("condition", "complete", "fraction", NULL),
  filt_thr = 3,
  filt_min = NULL
)
}
\arguments{
\item{data}{File or dataframe containing transcriptomics data, if \code{files.ind} is not used.}

\item{files.ind}{Prefixes of several files in the working directory containing transcriptomics data, if \code{data} is not used.}

\item{expdesign}{Experimental design as file path or data frame, if made previously.}

\item{csvsep}{Delimiter if reading CSV file(s).}

\item{dec}{Decimal separator if reading CSV, TSV, or TXT files.}

\item{sheet}{Sheet of an Excel file to be read.}

\item{name}{Header of column containing primary gene IDs (e.g., gene names).}

\item{id}{Header of column containing alternative gene IDs (e.g., locus IDs).}

\item{values}{Name of the column containing the values (if \code{files.ind != NULL}).}

\item{id2name.table}{File containing a table with ID to name mappings.}

\item{id2name.id}{Header of column containing alternative gene IDs in \code{id2name.table}.}

\item{id2name.name}{Header of column containing primary gene IDs in \code{id2name.table}.}

\item{pfx.counts}{Prefix in headers of columns containing gene abundances (if \code{data != NULL}).}

\item{rsd_thresh}{Provide a relative standard deviation (RSD) threshold \strong{in \%} for genes. The RSD is calculated for each condition and if the maximum RSD value determined for a given protein exceeds \code{rsd_thresh}, the gene is discarded. The RSD filter is applied \strong{before} further missing value filters based on the three \code{filt_} arguments.}

\item{filt_type}{(Character string) "complete", "condition" or "fraction", Sets the type of filtering applied. "complete" will only keep genes with valid values in all samples. "condition" will keep genes that have a maximum of \code{filt_thr} missing values in at least one condition. "fraction" will keep genes that have a \code{filt_min} fraction of valid values in all samples.}

\item{filt_thr}{(Integer) Sets the threshold for the allowed number of missing values in at least one condition if \code{filt_type = "condition"}. In other words: "keep genes that have a maximum of 'filt_thr' missing values in at least one condition."}

\item{filt_min}{(Numeric) Sets the threshold for the minimum fraction of valid values allowed for any genes if \code{filt_type = "fraction"}.}
}
\value{
Returns a \code{SummarizedExperiment} object.
}
\description{
Reads transcriptomics data from one/several file(s) or dataframe and creates a \code{SummarizedExperiment} object.
}
