% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcriptomics_computation.R
\name{rna.workflow}
\alias{rna.workflow}
\title{import SummarizedExperiment
import IHW}
\usage{
rna.workflow(
  se,
  imp_fun = c("zero", "man", "bpca", "knn", "QRILC", "MLE", "MinDet", "MinProb", "min",
    "zero", "mixed", "nbavg", "SampMin"),
  q = 0.01,
  knn.rowmax = 0.5,
  type = c("all", "control", "manual"),
  design = "~ condition",
  size.factors = NULL,
  altHypothesis = NULL,
  control = NULL,
  contrast = NULL,
  controlGenes = NULL,
  pAdjustMethod = c("IHW", "BH"),
  alpha = 0.05,
  alpha.independent = 0.1,
  alpha_pathways = 0.1,
  lfcShrink = TRUE,
  shrink.method = c("apeglm", "ashr", "normal"),
  lfc = 2,
  heatmap.show_all = TRUE,
  heatmap.kmeans = F,
  k = 6,
  heatmap.col_limit = NA,
  heatmap.show_row_names = TRUE,
  heatmap.row_font_size = 6,
  volcano.add_names = FALSE,
  volcano.label_size = 2.5,
  volcano.adjusted = TRUE,
  plot = FALSE,
  export = FALSE,
  report = TRUE,
  report.dir = NULL,
  pathway_enrichment = FALSE,
  pathway_kegg = FALSE,
  kegg_organism = NULL,
  custom_pathways = NULL,
  quiet = FALSE
)
}
\description{
import SummarizedExperiment
import IHW
}
